Move parser into its own directory
