--- turo/lib/parser-generated.js
@@ -1,5254 +1,0 @@
-import helper from '../lib/parser-helper.js';
-import ast from "../lib/ast";
-
-export default (function(){
-  /*
-   * Generated by PEG.js 0.7.0.
-   *
-   * http://pegjs.majda.cz/
-   */
-  
-  function quote(s) {
-    /*
-     * ECMA-262, 5th ed., 7.8.4: All characters may appear literally in a
-     * string literal except for the closing quote character, backslash,
-     * carriage return, line separator, paragraph separator, and line feed.
-     * Any character may appear in the form of an escape sequence.
-     *
-     * For portability, we also escape escape all control and non-ASCII
-     * characters. Note that "\0" and "\v" escape sequences are not used
-     * because JSHint does not like the first and IE the second.
-     */
-     return '"' + s
-      .replace(/\\/g, '\\\\')  // backslash
-      .replace(/"/g, '\\"')    // closing quote character
-      .replace(/\x08/g, '\\b') // backspace
-      .replace(/\t/g, '\\t')   // horizontal tab
-      .replace(/\n/g, '\\n')   // line feed
-      .replace(/\f/g, '\\f')   // form feed
-      .replace(/\r/g, '\\r')   // carriage return
-      .replace(/[\x00-\x07\x0B\x0E-\x1F\x80-\uFFFF]/g, escape)
-      + '"';
-  }
-  
-  var result = {
-    /*
-     * Parses the input with a generated parser. If the parsing is successfull,
-     * returns a value explicitly or implicitly specified by the grammar from
-     * which the parser was generated (see |PEG.buildParser|). If the parsing is
-     * unsuccessful, throws |PEG.parser.SyntaxError| describing the error.
-     */
-    parse: function(input, startRule) {
-      var parseFunctions = {
-        "StatementList": parse_StatementList,
-        "StatementSeperator": parse_StatementSeperator,
-        "EditorText": parse_EditorText,
-        "EditorTextLine": parse_EditorTextLine,
-        "DocumentFirstParse": parse_DocumentFirstParse,
-        "DocumentLineFirstParse": parse_DocumentLineFirstParse,
-        "FirstParseStatement": parse_FirstParseStatement,
-        "FirstParseUnitDefinition": parse_FirstParseUnitDefinition,
-        "FirstParseVariableDefinition": parse_FirstParseVariableDefinition,
-        "RestOfLine": parse_RestOfLine,
-        "CR": parse_CR,
-        "PaddedStatement": parse_PaddedStatement,
-        "Statement": parse_Statement,
-        "RawStatement": parse_RawStatement,
-        "UnitDefinitionStatement": parse_UnitDefinitionStatement,
-        "UnitDefinition": parse_UnitDefinition,
-        "UnitSchemeAnnotation": parse_UnitSchemeAnnotation,
-        "NewUnitDefinition": parse_NewUnitDefinition,
-        "UnitAssignmentLiteral": parse_UnitAssignmentLiteral,
-        "UnitDefinitionLeft": parse_UnitDefinitionLeft,
-        "UnitIdentifiers": parse_UnitIdentifiers,
-        "AlternativeUnitNames": parse_AlternativeUnitNames,
-        "UnitSchemes": parse_UnitSchemes,
-        "OptionalDimensionIdentifier": parse_OptionalDimensionIdentifier,
-        "DimensionIdentifier": parse_DimensionIdentifier,
-        "NewUnitIdentifier": parse_NewUnitIdentifier,
-        "UnitScheme": parse_UnitScheme,
-        "UnitExpression": parse_UnitExpression,
-        "UnitMultiplicativeOperator": parse_UnitMultiplicativeOperator,
-        "UnitPer": parse_UnitPer,
-        "UnitPerLiteral": parse_UnitPerLiteral,
-        "UnitMultiplier": parse_UnitMultiplier,
-        "RaiseToPowerOperator": parse_RaiseToPowerOperator,
-        "RaiseToPowerOperatorLiteral": parse_RaiseToPowerOperatorLiteral,
-        "UnitIdentifier": parse_UnitIdentifier,
-        "IncludeStatement": parse_IncludeStatement,
-        "ImportStatement": parse_ImportStatement,
-        "StringLiteral": parse_StringLiteral,
-        "ConstDefinition": parse_ConstDefinition,
-        "VariableDefinition": parse_VariableDefinition,
-        "AssignmentLiteral": parse_AssignmentLiteral,
-        "NumericalExpression": parse_NumericalExpression,
-        "ComparatorLiteral": parse_ComparatorLiteral,
-        "NumbericalExpressionUnitConversion": parse_NumbericalExpressionUnitConversion,
-        "UnitConversion": parse_UnitConversion,
-        "UnitConversionNode": parse_UnitConversionNode,
-        "AdditiveExpression": parse_AdditiveExpression,
-        "AdditiveOperator": parse_AdditiveOperator,
-        "AdditiveOperatorLiteral": parse_AdditiveOperatorLiteral,
-        "MultiplicativeExpression": parse_MultiplicativeExpression,
-        "MultiplicativeOperator": parse_MultiplicativeOperator,
-        "MultiplicativeOperatorLiteral": parse_MultiplicativeOperatorLiteral,
-        "MultiplactionOperand": parse_MultiplactionOperand,
-        "MultiplactionOperandWithoutUnit": parse_MultiplactionOperandWithoutUnit,
-        "NamedBinaryOperation": parse_NamedBinaryOperation,
-        "NamedBinaryOperator": parse_NamedBinaryOperator,
-        "NamedBinaryOperatorLiteral": parse_NamedBinaryOperatorLiteral,
-        "SpecialNamedBinaryOperator": parse_SpecialNamedBinaryOperator,
-        "NamedPrefixUnaryOperation": parse_NamedPrefixUnaryOperation,
-        "NamedPrefixOperator": parse_NamedPrefixOperator,
-        "SymbolTablePrefixOperator": parse_SymbolTablePrefixOperator,
-        "NamedPostfixUnaryOperation": parse_NamedPostfixUnaryOperation,
-        "NamedPostfixOperator": parse_NamedPostfixOperator,
-        "SymbolTablePostfixOperator": parse_SymbolTablePostfixOperator,
-        "ValueOrParens": parse_ValueOrParens,
-        "ParensOpen": parse_ParensOpen,
-        "ParensClose": parse_ParensClose,
-        "Terminal": parse_Terminal,
-        "Identifier": parse_Identifier,
-        "UnitIdentifierLiteral": parse_UnitIdentifierLiteral,
-        "SpecialUnitLiteral": parse_SpecialUnitLiteral,
-        "IdentifierLiteral": parse_IdentifierLiteral,
-        "Number": parse_Number,
-        "NumberLiteral": parse_NumberLiteral,
-        "NumberLiteralString": parse_NumberLiteralString,
-        "NumberLiteralFromParts": parse_NumberLiteralFromParts,
-        "NumberWholePart": parse_NumberWholePart,
-        "Digits": parse_Digits,
-        "NumberDecimalPart": parse_NumberDecimalPart,
-        "DecimalPoint": parse_DecimalPoint,
-        "ExponentLiteral": parse_ExponentLiteral,
-        "ExponentSign": parse_ExponentSign,
-        "IntegerNode": parse_IntegerNode,
-        "IntegerLiteral": parse_IntegerLiteral,
-        "IntegerString": parse_IntegerString,
-        "KeywordLiteral": parse_KeywordLiteral,
-        "_": parse__,
-        "Comment": parse_Comment,
-        "__": parse___,
-        "nbsp": parse_nbsp,
-        "nbsp_1": parse_nbsp_1,
-        "IdentifierStart": parse_IdentifierStart,
-        "IdentifierPart": parse_IdentifierPart,
-        "UnicodeLetter": parse_UnicodeLetter,
-        "UnicodeCombiningMark": parse_UnicodeCombiningMark,
-        "UnicodeEscapeSequence": parse_UnicodeEscapeSequence,
-        "HexDigit": parse_HexDigit,
-        "Ll": parse_Ll,
-        "Lm": parse_Lm,
-        "Lo": parse_Lo,
-        "Lt": parse_Lt,
-        "Lu": parse_Lu,
-        "Mc": parse_Mc,
-        "Mn": parse_Mn,
-        "Nd": parse_Nd,
-        "Nl": parse_Nl,
-        "Pc": parse_Pc,
-        "Zs": parse_Zs
-      };
-      
-      if (startRule !== undefined) {
-        if (parseFunctions[startRule] === undefined) {
-          throw new Error("Invalid rule name: " + quote(startRule) + ".");
-        }
-      } else {
-        startRule = "PaddedStatement";
-      }
-      
-      var pos = { offset: 0, line: 1, column: 1, seenCR: false };
-      var reportFailures = 0;
-      var rightmostFailuresPos = { offset: 0, line: 1, column: 1, seenCR: false };
-      var rightmostFailuresExpected = [];
-      
-      function padLeft(input, padding, length) {
-        var result = input;
-        
-        var padLength = length - input.length;
-        for (var i = 0; i < padLength; i++) {
-          result = padding + result;
-        }
-        
-        return result;
-      }
-      
-      function escape(ch) {
-        var charCode = ch.charCodeAt(0);
-        var escapeChar;
-        var length;
-        
-        if (charCode <= 0xFF) {
-          escapeChar = 'x';
-          length = 2;
-        } else {
-          escapeChar = 'u';
-          length = 4;
-        }
-        
-        return '\\' + escapeChar + padLeft(charCode.toString(16).toUpperCase(), '0', length);
-      }
-      
-      function clone(object) {
-        var result = {};
-        for (var key in object) {
-          result[key] = object[key];
-        }
-        return result;
-      }
-      
-      function advance(pos, n) {
-        var endOffset = pos.offset + n;
-        
-        for (var offset = pos.offset; offset < endOffset; offset++) {
-          var ch = input.charAt(offset);
-          if (ch === "\n") {
-            if (!pos.seenCR) { pos.line++; }
-            pos.column = 1;
-            pos.seenCR = false;
-          } else if (ch === "\r" || ch === "\u2028" || ch === "\u2029") {
-            pos.line++;
-            pos.column = 1;
-            pos.seenCR = true;
-          } else {
-            pos.column++;
-            pos.seenCR = false;
-          }
-        }
-        
-        pos.offset += n;
-      }
-      
-      function matchFailed(failure) {
-        if (pos.offset < rightmostFailuresPos.offset) {
-          return;
-        }
-        
-        if (pos.offset > rightmostFailuresPos.offset) {
-          rightmostFailuresPos = clone(pos);
-          rightmostFailuresExpected = [];
-        }
-        
-        rightmostFailuresExpected.push(failure);
-      }
-      
-      function parse_StatementList() {
-        var result0, result1, result2, result3;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse__();
-        if (result0 !== null) {
-          result1 = [];
-          pos2 = clone(pos);
-          result2 = parse_Statement();
-          result2 = result2 !== null ? result2 : "";
-          if (result2 !== null) {
-            result3 = parse_StatementSeperator();
-            if (result3 !== null) {
-              result2 = [result2, result3];
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos2);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos2 = clone(pos);
-            result2 = parse_Statement();
-            result2 = result2 !== null ? result2 : "";
-            if (result2 !== null) {
-              result3 = parse_StatementSeperator();
-              if (result3 !== null) {
-                result2 = [result2, result3];
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, list) {
-          var statements = [];
-          for (var i=0, max = list.length; i<max; i+=2) {
-            if (list[i][0]) {
-              statements.push(list[i][0]);
-            }
-          }
-          return new ast.StatementNode("StatementList", statements);
-        })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_StatementSeperator() {
-        var result0, result1, result2;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse__();
-        if (result0 !== null) {
-          if (input.charCodeAt(pos.offset) === 59) {
-            result1 = ";";
-            advance(pos, 1);
-          } else {
-            result1 = null;
-            if (reportFailures === 0) {
-              matchFailed("\";\"");
-            }
-          }
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos0);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos0);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_EditorText() {
-        var result0, result1;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = [];
-        result1 = parse_EditorTextLine();
-        while (result1 !== null) {
-          result0.push(result1);
-          result1 = parse_EditorTextLine();
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, lines) {
-              return new ast.EditorLinesNode(lines);
-            })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_EditorTextLine() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_nbsp();
-        if (result0 !== null) {
-          result1 = parse_Statement();
-          if (result1 !== null) {
-            result2 = parse_RestOfLine();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, node, lineEnd) {
-              return helper.decorateStatement(node, line, offset, lineEnd);
-            })(pos0.offset, pos0.line, pos0.column, result0[1], result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          result0 = parse_RestOfLine();
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, lineEnd) {
-                return helper.decorateStatement(null, line, offset, lineEnd);
-              })(pos0.offset, pos0.line, pos0.column, result0);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_DocumentFirstParse() {
-        var result0, result1;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = [];
-        result1 = parse_DocumentLineFirstParse();
-        while (result1 !== null) {
-          result0.push(result1);
-          result1 = parse_DocumentLineFirstParse();
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, lines) {
-              return new ast.EditorLinesNode(lines);
-            })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_DocumentLineFirstParse() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_nbsp();
-        if (result0 !== null) {
-          result1 = parse_FirstParseStatement();
-          if (result1 !== null) {
-            result2 = parse_RestOfLine();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, node, lineEnd) {
-              return helper.decorateStatement(node, line, offset, lineEnd);
-            })(pos0.offset, pos0.line, pos0.column, result0[1], result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          result0 = parse_RestOfLine();
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, lineEnd) {
-                return helper.decorateStatement(null, line, offset, lineEnd);
-              })(pos0.offset, pos0.line, pos0.column, result0);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-
-      function parse_FirstParseStatement() {
-        var result0;
-
-        result0 = parse_ImportStatement();
-        if (result0 === null) {
-          result0 = parse_FirstParseUnitDefinition();
-          if (result0 === null) {
-            result0 = parse_FirstParseVariableDefinition();
-          }
-        }
-        return result0;
-      }
-
-      function parse_FirstParseUnitDefinition() {
-        var result0, result1, result2, result3, result4, result5;
-        var pos0, pos1, pos2;
-
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.substr(pos.offset, 4) === "unit") {
-          result0 = "unit";
-          advance(pos, 4);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"unit\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse___();
-          if (result1 !== null) {
-            pos2 = clone(pos);
-            result2 = parse_NumberLiteral();
-            if (result2 !== null) {
-              result3 = parse__();
-              if (result3 !== null) {
-                result2 = [result2, result3];
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-            result2 = result2 !== null ? result2 : "";
-            if (result2 !== null) {
-              result3 = parse_UnitDefinitionLeft();
-              if (result3 !== null) {
-                pos2 = clone(pos);
-                result4 = parse__();
-                if (result4 !== null) {
-                  result5 = parse_UnitAssignmentLiteral();
-                  if (result5 !== null) {
-                    result4 = [result4, result5];
-                  } else {
-                    result4 = null;
-                    pos = clone(pos2);
-                  }
-                } else {
-                  result4 = null;
-                  pos = clone(pos2);
-                }
-                result4 = result4 !== null ? result4 : "";
-                if (result4 !== null) {
-                  result0 = [result0, result1, result2, result3, result4];
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unit, colon) {
-              // unit = { id, alternatives, schemes }
-        
-              if (colon) {
-                // Brand new unit
-                parseContext.scope.addUnit(unit, unit.id, unit.alternatives);
-              } else {
-                // New alternative names for an existing unit
-                parseContext.scope.addUnit(unit, null, unit.alternatives);
-              }
-            })(pos0.offset, pos0.line, pos0.column, result0[3], result0[4]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_FirstParseVariableDefinition() {
-        var result0, result1, result2, result3;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        pos2 = clone(pos);
-        if (input.substr(pos.offset, 5) === "const") {
-          result0 = "const";
-          advance(pos, 5);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"const\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse___();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos2);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos2);
-        }
-        result0 = result0 !== null ? result0 : "";
-        if (result0 !== null) {
-          result1 = parse_IdentifierLiteral();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_AssignmentLiteral();
-              if (result3 !== null) {
-                result0 = [result0, result1, result2, result3];
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, id) {
-              parseContext.scope.addVariable(id, {});
-            })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_RestOfLine() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = [];
-        if (/^[^\r\n]/.test(input.charAt(pos.offset))) {
-          result1 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result1 = null;
-          if (reportFailures === 0) {
-            matchFailed("[^\\r\\n]");
-          }
-        }
-        while (result1 !== null) {
-          result0.push(result1);
-          if (/^[^\r\n]/.test(input.charAt(pos.offset))) {
-            result1 = input.charAt(pos.offset);
-            advance(pos, 1);
-          } else {
-            result1 = null;
-            if (reportFailures === 0) {
-              matchFailed("[^\\r\\n]");
-            }
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse_CR();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, text, eol) {
-              return {
-                text: text.join(''),
-                offsetLast: eol[0],
-                lineLast: eol[1],
-              };
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_CR() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        if (/^[\r\n]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\r\\n]");
-          }
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column) {
-              return [offset, line];
-            })(pos0.offset, pos0.line, pos0.column);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-
-      function parse_PaddedStatement() {
-        var result0, result1, result2;
-        var pos0, pos1;
-
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse__();
-        if (result0 !== null) {
-          result1 = parse_Statement();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, node) {
-            return node;
-          })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-
-      function parse_Statement() {
-        var result0;
-        var pos0;
-
-        pos0 = clone(pos);
-        result0 = parse_RawStatement();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, node) {
-            node.scope = parseContext.scope;
-            return node;
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-
-      function parse_RawStatement() {
-        var result0;
-
-        result0 = parse_UnitDefinitionStatement();
-        if (result0 === null) {
-          result0 = parse_IncludeStatement();
-          if (result0 === null) {
-            result0 = parse_ImportStatement();
-            if (result0 === null) {
-              result0 = parse_ConstDefinition();
-              if (result0 === null) {
-                result0 = parse_VariableDefinition();
-                if (result0 === null) {
-                  result0 = parse_NumericalExpression();
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitDefinitionStatement() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.substr(pos.offset, 4) === "unit") {
-          result0 = "unit";
-          advance(pos, 4);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"unit\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse___();
-          if (result1 !== null) {
-            result2 = parse_UnitDefinition();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unit) {
-          return new ast.StatementNode("UnitDefinition", unit);
-        })(pos0.offset, pos0.line, pos0.column, result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_UnitDefinition() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_NewUnitDefinition();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unit) {
-              parseContext.scope.addUnit(unit, unit.name, unit.alternatives);
-              return unit;
-            })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          result0 = parse_UnitSchemeAnnotation();
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, unit) {
-                return unit;
-              })(pos0.offset, pos0.line, pos0.column, result0);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitSchemeAnnotation() {
-        var result0, result1, result2, result3;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_UnitIdentifier();
-        if (result0 !== null) {
-          result1 = parse_AlternativeUnitNames();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_UnitSchemes();
-              result3 = result3 !== null ? result3 : "";
-              if (result3 !== null) {
-                result0 = [result0, result1, result2, result3];
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unitNode, alternatives, schemes) {
-            // Unsure if this the right thing to do: 
-            // allow both alternatives and schemes to be optional.
-            if (schemes) {
-              units.addUnitSchemes(unitNode.unit, schemes);
-            }
-            if (alternatives.length) {
-              units.addUnitAliases(unitNode.unit, alternatives);
-              parseContext.scope.addUnit(unitNode.unit, null, alternatives);
-            }
-            return unitNode.unit;
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1], result0[3]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NewUnitDefinition() {
-        var result0, result1, result2, result3, result4, result5, result6, result7, result8;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_UnitDefinitionLeft();
-        if (result0 !== null) {
-          result1 = parse__();
-          if (result1 !== null) {
-            result2 = parse_UnitAssignmentLiteral();
-            if (result2 !== null) {
-              result3 = parse__();
-              if (result3 !== null) {
-                result4 = parse_OptionalDimensionIdentifier();
-                result4 = result4 !== null ? result4 : "";
-                if (result4 !== null) {
-                  result5 = parse__();
-                  if (result5 !== null) {
-                    result6 = parse_NumberLiteral();
-                    if (result6 !== null) {
-                      result7 = parse__();
-                      if (result7 !== null) {
-                        result8 = parse_UnitExpression();
-                        if (result8 !== null) {
-                          result0 = [result0, result1, result2, result3, result4, result5, result6, result7, result8];
-                        } else {
-                          result0 = null;
-                          pos = clone(pos1);
-                        }
-                      } else {
-                        result0 = null;
-                        pos = clone(pos1);
-                      }
-                    } else {
-                      result0 = null;
-                      pos = clone(pos1);
-                    }
-                  } else {
-                    result0 = null;
-                    pos = clone(pos1);
-                  }
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, id, dimension, num, unitNode) {
-              return units.addUnit(id.id, num || 1.0, unitNode.unit, id.schemes, dimension, id.alternatives);
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[4], result0[6], result0[8]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          pos1 = clone(pos);
-          result0 = parse_UnitDefinitionLeft();
-          if (result0 !== null) {
-            result1 = parse__();
-            if (result1 !== null) {
-              result2 = parse_UnitAssignmentLiteral();
-              if (result2 !== null) {
-                result3 = parse__();
-                if (result3 !== null) {
-                  result4 = parse_DimensionIdentifier();
-                  if (result4 !== null) {
-                    result0 = [result0, result1, result2, result3, result4];
-                  } else {
-                    result0 = null;
-                    pos = clone(pos1);
-                  }
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, id, dimension) {
-                return units.addUnit(id.id, dimension, undefined, id.schemes, undefined, id.alternatives);
-              })(pos0.offset, pos0.line, pos0.column, result0[0], result0[4]);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-          if (result0 === null) {
-            pos0 = clone(pos);
-            pos1 = clone(pos);
-            result0 = parse_NumberLiteral();
-            if (result0 !== null) {
-              result1 = parse__();
-              if (result1 !== null) {
-                result2 = parse_UnitDefinitionLeft();
-                if (result2 !== null) {
-                  result3 = parse__();
-                  if (result3 !== null) {
-                    result4 = parse_UnitAssignmentLiteral();
-                    if (result4 !== null) {
-                      result5 = parse__();
-                      if (result5 !== null) {
-                        result6 = parse_OptionalDimensionIdentifier();
-                        result6 = result6 !== null ? result6 : "";
-                        if (result6 !== null) {
-                          result7 = parse__();
-                          if (result7 !== null) {
-                            result8 = parse_UnitExpression();
-                            if (result8 !== null) {
-                              result0 = [result0, result1, result2, result3, result4, result5, result6, result7, result8];
-                            } else {
-                              result0 = null;
-                              pos = clone(pos1);
-                            }
-                          } else {
-                            result0 = null;
-                            pos = clone(pos1);
-                          }
-                        } else {
-                          result0 = null;
-                          pos = clone(pos1);
-                        }
-                      } else {
-                        result0 = null;
-                        pos = clone(pos1);
-                      }
-                    } else {
-                      result0 = null;
-                      pos = clone(pos1);
-                    }
-                  } else {
-                    result0 = null;
-                    pos = clone(pos1);
-                  }
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-            if (result0 !== null) {
-              result0 = (function(offset, line, column, num, id, dimension, unitNode) {
-                  return units.addUnit(id.id, unitNode.unit, num, id.schemes, dimension, id.alternatives);
-                })(pos0.offset, pos0.line, pos0.column, result0[0], result0[2], result0[6], result0[8]);
-            }
-            if (result0 === null) {
-              pos = clone(pos0);
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitAssignmentLiteral() {
-        var result0;
-        
-        result0 = parse_AssignmentLiteral();
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 58) {
-            result0 = ":";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\":\"");
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitDefinitionLeft() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_UnitIdentifiers();
-        if (result0 !== null) {
-          result1 = parse__();
-          if (result1 !== null) {
-            result2 = parse_UnitSchemes();
-            result2 = result2 !== null ? result2 : "";
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, ids, schemes) {
-            return {
-              id: ids[0],
-              alternatives: ids[1],
-              schemes: schemes
-            }
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_UnitIdentifiers() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_NewUnitIdentifier();
-        if (result0 !== null) {
-          result1 = parse_AlternativeUnitNames();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unitName, alternatives) {
-          return [unitName, alternatives];
-        })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_AlternativeUnitNames() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        result0 = [];
-        pos1 = clone(pos);
-        result1 = parse___();
-        if (result1 !== null) {
-          result2 = parse_NewUnitIdentifier();
-          if (result2 !== null) {
-            result1 = [result1, result2];
-          } else {
-            result1 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result1 = null;
-          pos = clone(pos1);
-        }
-        while (result1 !== null) {
-          result0.push(result1);
-          pos1 = clone(pos);
-          result1 = parse___();
-          if (result1 !== null) {
-            result2 = parse_NewUnitIdentifier();
-            if (result2 !== null) {
-              result1 = [result1, result2];
-            } else {
-              result1 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result1 = null;
-            pos = clone(pos1);
-          }
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, alternatives) {
-          var alts = [];
-          for (var i=0, max=alternatives.length; i<max; i++) {
-            alts.push(alternatives[i][1]);
-          }
-          return alts;
-        })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_UnitSchemes() {
-        var result0, result1, result2, result3;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_ParensOpen();
-        if (result0 !== null) {
-          result1 = [];
-          pos2 = clone(pos);
-          result2 = parse__();
-          if (result2 !== null) {
-            result3 = parse_UnitScheme();
-            if (result3 !== null) {
-              result2 = [result2, result3];
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos2);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos2 = clone(pos);
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_UnitScheme();
-              if (result3 !== null) {
-                result2 = [result2, result3];
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          }
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_ParensClose();
-              if (result3 !== null) {
-                result0 = [result0, result1, result2, result3];
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, schemeLiterals) {
-          if (!schemeLiterals.length) {
-            return;
-          }
-          var schemes = [];
-          for (var i=0, max=schemeLiterals.length; i<max; i++) {
-            schemes.push(schemeLiterals[i][1]);
-          }
-        
-          return schemes;
-        })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_OptionalDimensionIdentifier() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_DimensionIdentifier();
-        if (result0 !== null) {
-          result1 = parse__();
-          if (result1 !== null) {
-            if (input.charCodeAt(pos.offset) === 44) {
-              result2 = ",";
-              advance(pos, 1);
-            } else {
-              result2 = null;
-              if (reportFailures === 0) {
-                matchFailed("\",\"");
-              }
-            }
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, dimension) {
-          return dimension;
-        })(pos0.offset, pos0.line, pos0.column, result0[0]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_DimensionIdentifier() {
-        var result0;
-        
-        reportFailures++;
-        result0 = parse_IdentifierLiteral();
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("dimensionName");
-        }
-        return result0;
-      }
-      
-      function parse_NewUnitIdentifier() {
-        var result0;
-        
-        reportFailures++;
-        result0 = parse_UnitIdentifierLiteral();
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("newUnitName");
-        }
-        return result0;
-      }
-      
-      function parse_UnitScheme() {
-        var result0;
-        
-        reportFailures++;
-        result0 = parse_IdentifierLiteral();
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("unitScheme");
-        }
-        return result0;
-      }
-      
-      function parse_UnitExpression() {
-        var result0, result1, result2, result3, result4, result5;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_nbsp();
-        if (result0 !== null) {
-          pos2 = clone(pos);
-          result1 = parse_UnitPer();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result1 = [result1, result2];
-            } else {
-              result1 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-          result1 = result1 !== null ? result1 : "";
-          if (result1 !== null) {
-            result2 = parse_UnitMultiplier();
-            if (result2 !== null) {
-              result3 = [];
-              pos2 = clone(pos);
-              result4 = parse_UnitMultiplicativeOperator();
-              if (result4 !== null) {
-                result5 = parse_UnitMultiplier();
-                if (result5 !== null) {
-                  result4 = [result4, result5];
-                } else {
-                  result4 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result4 = null;
-                pos = clone(pos2);
-              }
-              while (result4 !== null) {
-                result3.push(result4);
-                pos2 = clone(pos);
-                result4 = parse_UnitMultiplicativeOperator();
-                if (result4 !== null) {
-                  result5 = parse_UnitMultiplier();
-                  if (result5 !== null) {
-                    result4 = [result4, result5];
-                  } else {
-                    result4 = null;
-                    pos = clone(pos2);
-                  }
-                } else {
-                  result4 = null;
-                  pos = clone(pos2);
-                }
-              }
-              if (result3 !== null) {
-                result0 = [result0, result1, result2, result3];
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, optionalPer, head, tail) {
-            var unitNode = head,
-                op;
-            for (var i=0; i<tail.length; i++) {
-              op = tail[i][0];
-              unitNode = new ast.UnitMultOp(helper.infixAliases[op.opName] || op.opName, unitNode, tail[i][1]);
-              helper.decorateNonTerminal(unitNode, op.offset, op.literal);
-            }
-            
-            op = optionalPer[0];
-            if (op) {
-              unitNode = new ast.UnitMultOp('/', null, unitNode);
-              unitNode._offsetFirst = offset;
-              helper.decorateNonTerminal(unitNode, op.offset, op.literal);
-            }
-            return unitNode;
-        })(pos0.offset, pos0.line, pos0.column, result0[1], result0[2], result0[3]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_UnitMultiplicativeOperator() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse__();
-        if (result0 !== null) {
-          result1 = parse_UnitPer();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, op) { return op; })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          result0 = parse_nbsp_1();
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, space) { 
-                return {
-                  offset: offset,
-                  literal: space,
-                  opName: '*',
-                };
-              })(pos0.offset, pos0.line, pos0.column, result0);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitPer() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = parse_UnitPerLiteral();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-            return {
-              offset: offset,
-              literal: literal,
-              opName: '/'
-            };
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("unitPer");
-        }
-        return result0;
-      }
-      
-      function parse_UnitPerLiteral() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        if (input.substr(pos.offset, 3) === "per") {
-          result0 = "per";
-          advance(pos, 3);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"per\"");
-          }
-        }
-        if (result0 !== null) {
-          pos1 = clone(pos);
-          reportFailures++;
-          result1 = parse_IdentifierPart();
-          reportFailures--;
-          if (result1 === null) {
-            result1 = "";
-          } else {
-            result1 = null;
-            pos = clone(pos1);
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos0);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 47) {
-            result0 = "/";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"/\"");
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitMultiplier() {
-        var result0, result1, result2, result3, result4;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_UnitIdentifier();
-        if (result0 !== null) {
-          result1 = parse__();
-          if (result1 !== null) {
-            result2 = parse_RaiseToPowerOperator();
-            if (result2 !== null) {
-              result3 = parse__();
-              if (result3 !== null) {
-                result4 = parse_IntegerNode();
-                if (result4 !== null) {
-                  result0 = [result0, result1, result2, result3, result4];
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unitNode, op, num) { 
-              var node = new ast.UnitPowerNode(unitNode, num); 
-              return helper.decorateNonTerminal(node, op.offset, op.literal);
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[2], result0[4]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          result0 = parse_UnitIdentifier();
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, unitNode) {
-                return unitNode;
-              })(pos0.offset, pos0.line, pos0.column, result0);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_RaiseToPowerOperator() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = parse_RaiseToPowerOperatorLiteral();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-            return {
-              literal: literal,
-              offset: offset,
-            }
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("unitPower");
-        }
-        return result0;
-      }
-      
-      function parse_RaiseToPowerOperatorLiteral() {
-        var result0;
-        var pos0;
-        
-        if (input.charCodeAt(pos.offset) === 94) {
-          result0 = "^";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"^\"");
-          }
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          if (input.substr(pos.offset, 2) === "**") {
-            result0 = "**";
-            advance(pos, 2);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"**\"");
-            }
-          }
-          if (result0 !== null) {
-            result0 = (function(offset, line, column) { return "^"; })(pos0.offset, pos0.line, pos0.column);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnitIdentifier() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_UnitIdentifierLiteral();
-        if (result0 !== null) {
-          result1 = (function(offset, line, column, unitName) { return !!parseContext.scope.findUnit(unitName); })(pos.offset, pos.line, pos.column, result0) ? "" : null;
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, unitName) {
-          var node = new ast.UnitLiteralNode(parseContext.scope.findUnit(unitName), unitName);
-          return helper.decorateTerminal(node, offset, unitName);
-         })(pos0.offset, pos0.line, pos0.column, result0[0]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("unit");
-        }
-        return result0;
-      }
-      
-      function parse_IncludeStatement() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.substr(pos.offset, 7) === "include") {
-          result0 = "include";
-          advance(pos, 7);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"include\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse__();
-          if (result1 !== null) {
-            result2 = parse_StringLiteral();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, filename) {
-          turo.include(filename);
-          return new ast.StatementNode("Include", filename);
-        })(pos0.offset, pos0.line, pos0.column, result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_ImportStatement() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.substr(pos.offset, 6) === "import") {
-          result0 = "import";
-          advance(pos, 6);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"import\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse___();
-          if (result1 !== null) {
-            result2 = parse_StringLiteral();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, filename) {
-          return new ast.StatementNode("Import", filename);
-        })(pos0.offset, pos0.line, pos0.column, result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_StringLiteral() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.charCodeAt(pos.offset) === 34) {
-          result0 = "\"";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"\\\"\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = [];
-          if (/^[^"]/.test(input.charAt(pos.offset))) {
-            result2 = input.charAt(pos.offset);
-            advance(pos, 1);
-          } else {
-            result2 = null;
-            if (reportFailures === 0) {
-              matchFailed("[^\"]");
-            }
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            if (/^[^"]/.test(input.charAt(pos.offset))) {
-              result2 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result2 = null;
-              if (reportFailures === 0) {
-                matchFailed("[^\"]");
-              }
-            }
-          }
-          if (result1 !== null) {
-            if (input.charCodeAt(pos.offset) === 34) {
-              result2 = "\"";
-              advance(pos, 1);
-            } else {
-              result2 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"\\\"\"");
-              }
-            }
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, string) {
-          return string.join("");
-        })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_ConstDefinition() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.substr(pos.offset, 5) === "const") {
-          result0 = "const";
-          advance(pos, 5);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"const\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse___();
-          if (result1 !== null) {
-            result2 = parse_VariableDefinition();
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, variable) {
-          variable.isConstant = true;
-          return variable;
-        })(pos0.offset, pos0.line, pos0.column, result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_VariableDefinition() {
-        var result0, result1, result2, result3, result4;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_IdentifierLiteral();
-        if (result0 !== null) {
-          result1 = parse__();
-          if (result1 !== null) {
-            result2 = parse_AssignmentLiteral();
-            if (result2 !== null) {
-              result3 = parse__();
-              if (result3 !== null) {
-                result4 = parse_NumericalExpression();
-                if (result4 !== null) {
-                  result0 = [result0, result1, result2, result3, result4];
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, identifier, definition) {
-          return parseContext.scope.addVariable(identifier, new ast.VariableDefinition(identifier, definition));
-        })(pos0.offset, pos0.line, pos0.column, result0[0], result0[4]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_AssignmentLiteral() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        if (input.charCodeAt(pos.offset) === 58) {
-          result0 = ":";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\":\"");
-          }
-        }
-        result0 = result0 !== null ? result0 : "";
-        if (result0 !== null) {
-          if (input.charCodeAt(pos.offset) === 61) {
-            result1 = "=";
-            advance(pos, 1);
-          } else {
-            result1 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"=\"");
-            }
-          }
-          if (result1 !== null) {
-            pos1 = clone(pos);
-            reportFailures++;
-            if (input.charCodeAt(pos.offset) === 61) {
-              result2 = "=";
-              advance(pos, 1);
-            } else {
-              result2 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"=\"");
-              }
-            }
-            reportFailures--;
-            if (result2 === null) {
-              result2 = "";
-            } else {
-              result2 = null;
-              pos = clone(pos1);
-            }
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos0);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos0);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("assignment");
-        }
-        return result0;
-      }
-      
-      function parse_NumericalExpression() {
-        var result0, result1, result2, result3, result4;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_NumbericalExpressionUnitConversion();
-        if (result0 !== null) {
-          pos2 = clone(pos);
-          result1 = parse__();
-          if (result1 !== null) {
-            result2 = parse_ComparatorLiteral();
-            if (result2 !== null) {
-              result3 = parse__();
-              if (result3 !== null) {
-                result4 = parse_NumbericalExpressionUnitConversion();
-                if (result4 !== null) {
-                  result1 = [result1, result2, result3, result4];
-                } else {
-                  result1 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result1 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result1 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-          result1 = result1 !== null ? result1 : "";
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, lhs, operation) {
-            if (operation) {
-              return new ast.BinaryNode(lhs, operation[3], operation[1]);
-            }
-            return lhs;
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_ComparatorLiteral() {
-        var result0;
-        
-        reportFailures++;
-        if (input.substr(pos.offset, 2) === "!=") {
-          result0 = "!=";
-          advance(pos, 2);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"!=\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.substr(pos.offset, 2) === "<=") {
-            result0 = "<=";
-            advance(pos, 2);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"<=\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.substr(pos.offset, 2) === "==") {
-              result0 = "==";
-              advance(pos, 2);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"==\"");
-              }
-            }
-            if (result0 === null) {
-              if (input.substr(pos.offset, 2) === ">=") {
-                result0 = ">=";
-                advance(pos, 2);
-              } else {
-                result0 = null;
-                if (reportFailures === 0) {
-                  matchFailed("\">=\"");
-                }
-              }
-              if (result0 === null) {
-                if (input.charCodeAt(pos.offset) === 60) {
-                  result0 = "<";
-                  advance(pos, 1);
-                } else {
-                  result0 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("\"<\"");
-                  }
-                }
-                if (result0 === null) {
-                  if (input.charCodeAt(pos.offset) === 62) {
-                    result0 = ">";
-                    advance(pos, 1);
-                  } else {
-                    result0 = null;
-                    if (reportFailures === 0) {
-                      matchFailed("\">\"");
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("comparator");
-        }
-        return result0;
-      }
-      
-      function parse_NumbericalExpressionUnitConversion() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_AdditiveExpression();
-        if (result0 !== null) {
-          result1 = parse_UnitConversion();
-          result1 = result1 !== null ? result1 : "";
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, node, unitConversion) {
-            if (unitConversion) {
-              var op = unitConversion[0],
-                  unitNode = unitConversion[1];
-              node = new ast.BinaryNode(node, unitNode, op.literal);
-              helper.decorateNonTerminal(node, op.offset, op.literal); 
-            }
-            return node;
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_UnitConversion() {
-        var result0, result1, result2, result3;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse__();
-        if (result0 !== null) {
-          result1 = parse_UnitConversionNode();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_UnitExpression();
-              if (result3 !== null) {
-                result0 = [result0, result1, result2, result3];
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, op, unitNode) {
-          return [op, unitNode];
-        })(pos0.offset, pos0.line, pos0.column, result0[1], result0[3]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_UnitConversionNode() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        if (input.substr(pos.offset, 2) === "in") {
-          result0 = "in";
-          advance(pos, 2);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"in\"");
-          }
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column) {
-            return {
-              literal: 'in',
-              offset: offset,
-            };
-          })(pos0.offset, pos0.line, pos0.column);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("unitIn");
-        }
-        return result0;
-      }
-      
-      function parse_AdditiveExpression() {
-        var result0, result1, result2, result3, result4, result5;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_MultiplicativeExpression();
-        if (result0 !== null) {
-          result1 = [];
-          pos2 = clone(pos);
-          result2 = parse__();
-          if (result2 !== null) {
-            result3 = parse_AdditiveOperator();
-            if (result3 !== null) {
-              result4 = parse__();
-              if (result4 !== null) {
-                result5 = parse_MultiplicativeExpression();
-                if (result5 !== null) {
-                  result2 = [result2, result3, result4, result5];
-                } else {
-                  result2 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos2);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos2 = clone(pos);
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_AdditiveOperator();
-              if (result3 !== null) {
-                result4 = parse__();
-                if (result4 !== null) {
-                  result5 = parse_MultiplicativeExpression();
-                  if (result5 !== null) {
-                    result2 = [result2, result3, result4, result5];
-                  } else {
-                    result2 = null;
-                    pos = clone(pos2);
-                  }
-                } else {
-                  result2 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, head, tail) {
-              return helper.unpackLeftBinaryOperations(head, tail);
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_AdditiveOperator() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_AdditiveOperatorLiteral();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-            return {
-              literal: literal,
-              offset: offset
-            };
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_AdditiveOperatorLiteral() {
-        var result0;
-        
-        reportFailures++;
-        if (input.charCodeAt(pos.offset) === 43) {
-          result0 = "+";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"+\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 45) {
-            result0 = "-";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"-\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.substr(pos.offset, 2) === "||") {
-              result0 = "||";
-              advance(pos, 2);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"||\"");
-              }
-            }
-          }
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("plusMinus");
-        }
-        return result0;
-      }
-      
-      function parse_MultiplicativeExpression() {
-        var result0, result1, result2, result3, result4, result5;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_MultiplactionOperand();
-        if (result0 !== null) {
-          result1 = [];
-          pos2 = clone(pos);
-          result2 = parse__();
-          if (result2 !== null) {
-            result3 = parse_MultiplicativeOperator();
-            if (result3 !== null) {
-              result4 = parse__();
-              if (result4 !== null) {
-                result5 = parse_MultiplactionOperand();
-                if (result5 !== null) {
-                  result2 = [result2, result3, result4, result5];
-                } else {
-                  result2 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos2);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos2 = clone(pos);
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_MultiplicativeOperator();
-              if (result3 !== null) {
-                result4 = parse__();
-                if (result4 !== null) {
-                  result5 = parse_MultiplactionOperand();
-                  if (result5 !== null) {
-                    result2 = [result2, result3, result4, result5];
-                  } else {
-                    result2 = null;
-                    pos = clone(pos2);
-                  }
-                } else {
-                  result2 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, head, tail) {
-              return helper.unpackLeftBinaryOperations(head, tail);
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_MultiplicativeOperator() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = parse_MultiplicativeOperatorLiteral();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-            if (typeof literal !== 'string') {
-              literal = literal[0];
-            }
-            return {
-              literal: literal, 
-              offset: offset
-            };
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("multiplyDivide");
-        }
-        return result0;
-      }
-      
-      function parse_MultiplicativeOperatorLiteral() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        if (input.charCodeAt(pos.offset) === 42) {
-          result0 = "*";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"*\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 215) {
-            result0 = "\xD7";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"\\xD7\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.charCodeAt(pos.offset) === 47) {
-              result0 = "/";
-              advance(pos, 1);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"/\"");
-              }
-            }
-            if (result0 === null) {
-              if (input.charCodeAt(pos.offset) === 247) {
-                result0 = "\xF7";
-                advance(pos, 1);
-              } else {
-                result0 = null;
-                if (reportFailures === 0) {
-                  matchFailed("\"\\xF7\"");
-                }
-              }
-              if (result0 === null) {
-                if (input.substr(pos.offset, 2) === "&&") {
-                  result0 = "&&";
-                  advance(pos, 2);
-                } else {
-                  result0 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("\"&&\"");
-                  }
-                }
-                if (result0 === null) {
-                  pos0 = clone(pos);
-                  if (input.substr(pos.offset, 2) === "at") {
-                    result0 = "at";
-                    advance(pos, 2);
-                  } else {
-                    result0 = null;
-                    if (reportFailures === 0) {
-                      matchFailed("\"at\"");
-                    }
-                  }
-                  if (result0 !== null) {
-                    pos1 = clone(pos);
-                    reportFailures++;
-                    result1 = parse_IdentifierPart();
-                    reportFailures--;
-                    if (result1 === null) {
-                      result1 = "";
-                    } else {
-                      result1 = null;
-                      pos = clone(pos1);
-                    }
-                    if (result1 !== null) {
-                      result0 = [result0, result1];
-                    } else {
-                      result0 = null;
-                      pos = clone(pos0);
-                    }
-                  } else {
-                    result0 = null;
-                    pos = clone(pos0);
-                  }
-                  if (result0 === null) {
-                    pos0 = clone(pos);
-                    if (input.substr(pos.offset, 2) === "of") {
-                      result0 = "of";
-                      advance(pos, 2);
-                    } else {
-                      result0 = null;
-                      if (reportFailures === 0) {
-                        matchFailed("\"of\"");
-                      }
-                    }
-                    if (result0 !== null) {
-                      pos1 = clone(pos);
-                      reportFailures++;
-                      result1 = parse_IdentifierPart();
-                      reportFailures--;
-                      if (result1 === null) {
-                        result1 = "";
-                      } else {
-                        result1 = null;
-                        pos = clone(pos1);
-                      }
-                      if (result1 !== null) {
-                        result0 = [result0, result1];
-                      } else {
-                        result0 = null;
-                        pos = clone(pos0);
-                      }
-                    } else {
-                      result0 = null;
-                      pos = clone(pos0);
-                    }
-                    if (result0 === null) {
-                      pos0 = clone(pos);
-                      if (input.substr(pos.offset, 2) === "by") {
-                        result0 = "by";
-                        advance(pos, 2);
-                      } else {
-                        result0 = null;
-                        if (reportFailures === 0) {
-                          matchFailed("\"by\"");
-                        }
-                      }
-                      if (result0 !== null) {
-                        pos1 = clone(pos);
-                        reportFailures++;
-                        result1 = parse_IdentifierPart();
-                        reportFailures--;
-                        if (result1 === null) {
-                          result1 = "";
-                        } else {
-                          result1 = null;
-                          pos = clone(pos1);
-                        }
-                        if (result1 !== null) {
-                          result0 = [result0, result1];
-                        } else {
-                          result0 = null;
-                          pos = clone(pos0);
-                        }
-                      } else {
-                        result0 = null;
-                        pos = clone(pos0);
-                      }
-                      if (result0 === null) {
-                        pos0 = clone(pos);
-                        if (input.substr(pos.offset, 5) === "every") {
-                          result0 = "every";
-                          advance(pos, 5);
-                        } else {
-                          result0 = null;
-                          if (reportFailures === 0) {
-                            matchFailed("\"every\"");
-                          }
-                        }
-                        if (result0 !== null) {
-                          pos1 = clone(pos);
-                          reportFailures++;
-                          result1 = parse_IdentifierPart();
-                          reportFailures--;
-                          if (result1 === null) {
-                            result1 = "";
-                          } else {
-                            result1 = null;
-                            pos = clone(pos1);
-                          }
-                          if (result1 !== null) {
-                            result0 = [result0, result1];
-                          } else {
-                            result0 = null;
-                            pos = clone(pos0);
-                          }
-                        } else {
-                          result0 = null;
-                          pos = clone(pos0);
-                        }
-                        if (result0 === null) {
-                          pos0 = clone(pos);
-                          if (input.substr(pos.offset, 3) === "per") {
-                            result0 = "per";
-                            advance(pos, 3);
-                          } else {
-                            result0 = null;
-                            if (reportFailures === 0) {
-                              matchFailed("\"per\"");
-                            }
-                          }
-                          if (result0 !== null) {
-                            pos1 = clone(pos);
-                            reportFailures++;
-                            result1 = parse_IdentifierPart();
-                            reportFailures--;
-                            if (result1 === null) {
-                              result1 = "";
-                            } else {
-                              result1 = null;
-                              pos = clone(pos1);
-                            }
-                            if (result1 !== null) {
-                              result0 = [result0, result1];
-                            } else {
-                              result0 = null;
-                              pos = clone(pos0);
-                            }
-                          } else {
-                            result0 = null;
-                            pos = clone(pos0);
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_MultiplactionOperand() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_MultiplactionOperandWithoutUnit();
-        if (result0 !== null) {
-          result1 = parse_UnitExpression();
-          result1 = result1 !== null ? result1 : "";
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, node, unitNode) {
-              if (unitNode) {
-                node.unitLiteral = unitNode.unit;
-                node.unitNode = unitNode;
-                unitNode.valueNode = node;
-                delete node._offsetLast;
-              }
-              return node;
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_MultiplactionOperandWithoutUnit() {
-        var result0, result1, result2;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = [];
-        pos2 = clone(pos);
-        result1 = parse_AdditiveOperator();
-        if (result1 !== null) {
-          result2 = parse__();
-          if (result2 !== null) {
-            result1 = [result1, result2];
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-        } else {
-          result1 = null;
-          pos = clone(pos2);
-        }
-        while (result1 !== null) {
-          result0.push(result1);
-          pos2 = clone(pos);
-          result1 = parse_AdditiveOperator();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result1 = [result1, result2];
-            } else {
-              result1 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse_NamedBinaryOperation();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, prefixes, operand) {
-            return helper.unpackUnaryOperations(operand, prefixes, true);
-           })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NamedBinaryOperation() {
-        var result0, result1, result2, result3, result4, result5;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_NamedPostfixUnaryOperation();
-        if (result0 !== null) {
-          result1 = [];
-          pos2 = clone(pos);
-          result2 = parse__();
-          if (result2 !== null) {
-            result3 = parse_NamedBinaryOperator();
-            if (result3 !== null) {
-              result4 = parse__();
-              if (result4 !== null) {
-                result5 = parse_NamedPostfixUnaryOperation();
-                if (result5 !== null) {
-                  result2 = [result2, result3, result4, result5];
-                } else {
-                  result2 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos2);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos2 = clone(pos);
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_NamedBinaryOperator();
-              if (result3 !== null) {
-                result4 = parse__();
-                if (result4 !== null) {
-                  result5 = parse_NamedPostfixUnaryOperation();
-                  if (result5 !== null) {
-                    result2 = [result2, result3, result4, result5];
-                  } else {
-                    result2 = null;
-                    pos = clone(pos2);
-                  }
-                } else {
-                  result2 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, head, tail) {
-            return helper.unpackRightBinaryOperations(head, tail);
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NamedBinaryOperator() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = parse_NamedBinaryOperatorLiteral();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-            return {
-              literal: literal,
-              offset: offset,
-            }
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("infixOp");
-        }
-        return result0;
-      }
-      
-      function parse_NamedBinaryOperatorLiteral() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        result0 = parse_SpecialNamedBinaryOperator();
-        if (result0 === null) {
-          pos0 = clone(pos);
-          pos1 = clone(pos);
-          result0 = parse_IdentifierLiteral();
-          if (result0 !== null) {
-            result1 = (function(offset, line, column, operator) {
-                  return (operators && operators.hasInfixOperator(operator));
-                })(pos.offset, pos.line, pos.column, result0) ? "" : null;
-            if (result1 !== null) {
-              result0 = [result0, result1];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, operator) {
-                return operator;
-              })(pos0.offset, pos0.line, pos0.column, result0[0]);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_SpecialNamedBinaryOperator() {
-        var result0;
-        
-        if (input.charCodeAt(pos.offset) === 94) {
-          result0 = "^";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"^\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 8730) {
-            result0 = "\u221A";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"\\u221A\"");
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_NamedPrefixUnaryOperation() {
-        var result0, result1, result2;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = [];
-        pos2 = clone(pos);
-        result1 = parse_NamedPrefixOperator();
-        if (result1 !== null) {
-          result2 = parse__();
-          if (result2 !== null) {
-            result1 = [result1, result2];
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-        } else {
-          result1 = null;
-          pos = clone(pos2);
-        }
-        while (result1 !== null) {
-          result0.push(result1);
-          pos2 = clone(pos);
-          result1 = parse_NamedPrefixOperator();
-          if (result1 !== null) {
-            result2 = parse__();
-            if (result2 !== null) {
-              result1 = [result1, result2];
-            } else {
-              result1 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-        }
-        if (result0 !== null) {
-          result1 = parse_ValueOrParens();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, prefixes, value) {
-            value = helper.unpackUnaryOperations(value, prefixes, true);
-            //UnaryOperationNode
-            return value;
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NamedPrefixOperator() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = parse_SymbolTablePrefixOperator();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-              return {
-                literal: literal,
-                offset: offset
-              };
-            })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("prefixOp");
-        }
-        return result0;
-      }
-      
-      function parse_SymbolTablePrefixOperator() {
-        var result0, result1;
-        var pos0, pos1, pos2;
-        
-        if (input.substr(pos.offset, 4) === "sqrt") {
-          result0 = "sqrt";
-          advance(pos, 4);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"sqrt\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 8730) {
-            result0 = "\u221A";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"\\u221A\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.charCodeAt(pos.offset) === 45) {
-              result0 = "-";
-              advance(pos, 1);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"-\"");
-              }
-            }
-            if (result0 === null) {
-              pos0 = clone(pos);
-              pos1 = clone(pos);
-              if (input.charCodeAt(pos.offset) === 33) {
-                result0 = "!";
-                advance(pos, 1);
-              } else {
-                result0 = null;
-                if (reportFailures === 0) {
-                  matchFailed("\"!\"");
-                }
-              }
-              if (result0 !== null) {
-                pos2 = clone(pos);
-                reportFailures++;
-                if (input.charCodeAt(pos.offset) === 61) {
-                  result1 = "=";
-                  advance(pos, 1);
-                } else {
-                  result1 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("\"=\"");
-                  }
-                }
-                reportFailures--;
-                if (result1 === null) {
-                  result1 = "";
-                } else {
-                  result1 = null;
-                  pos = clone(pos2);
-                }
-                if (result1 !== null) {
-                  result0 = [result0, result1];
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-              if (result0 !== null) {
-                result0 = (function(offset, line, column) { return "!"; })(pos0.offset, pos0.line, pos0.column);
-              }
-              if (result0 === null) {
-                pos = clone(pos0);
-              }
-              if (result0 === null) {
-                pos0 = clone(pos);
-                pos1 = clone(pos);
-                result0 = parse_IdentifierLiteral();
-                if (result0 !== null) {
-                  result1 = (function(offset, line, column, literal) {
-                        return (operators && operators.hasPrefixOperator(literal));
-                      })(pos.offset, pos.line, pos.column, result0) ? "" : null;
-                  if (result1 !== null) {
-                    result0 = [result0, result1];
-                  } else {
-                    result0 = null;
-                    pos = clone(pos1);
-                  }
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-                if (result0 !== null) {
-                  result0 = (function(offset, line, column, literal) {
-                      return literal;
-                    })(pos0.offset, pos0.line, pos0.column, result0[0]);
-                }
-                if (result0 === null) {
-                  pos = clone(pos0);
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_NamedPostfixUnaryOperation() {
-        var result0, result1, result2, result3;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_NamedPrefixUnaryOperation();
-        if (result0 !== null) {
-          result1 = [];
-          pos2 = clone(pos);
-          result2 = parse__();
-          if (result2 !== null) {
-            result3 = parse_NamedPostfixOperator();
-            if (result3 !== null) {
-              result2 = [result2, result3];
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos2);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos2 = clone(pos);
-            result2 = parse__();
-            if (result2 !== null) {
-              result3 = parse_NamedPostfixOperator();
-              if (result3 !== null) {
-                result2 = [result2, result3];
-              } else {
-                result2 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos2);
-            }
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, value, postfixes) {
-            value = helper.unpackUnaryOperations(value, postfixes, false);
-            //UnaryOperationNode
-            return value;
-          })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NamedPostfixOperator() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = parse_SymbolTablePostfixOperator();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, literal) {
-              return {
-                literal: literal,
-                offset: offset
-              };
-            })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("postfixOp");
-        }
-        return result0;
-      }
-      
-      function parse_SymbolTablePostfixOperator() {
-        var result0, result1;
-        var pos0, pos1, pos2;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.charCodeAt(pos.offset) === 33) {
-          result0 = "!";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"!\"");
-          }
-        }
-        if (result0 !== null) {
-          pos2 = clone(pos);
-          reportFailures++;
-          if (input.charCodeAt(pos.offset) === 61) {
-            result1 = "=";
-            advance(pos, 1);
-          } else {
-            result1 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"=\"");
-            }
-          }
-          reportFailures--;
-          if (result1 === null) {
-            result1 = "";
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column) {
-              return "!";
-            })(pos0.offset, pos0.line, pos0.column);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 37) {
-            result0 = "%";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"%\"");
-            }
-          }
-          if (result0 === null) {
-            pos0 = clone(pos);
-            result0 = parse_IdentifierLiteral();
-            if (result0 !== null) {
-              result1 = (function(offset, line, column, literal) {
-                    return (operators && operators.hasPostfixOperator(literal));
-                  })(pos.offset, pos.line, pos.column, result0) ? "" : null;
-              if (result1 !== null) {
-                result0 = [result0, result1];
-              } else {
-                result0 = null;
-                pos = clone(pos0);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos0);
-            }
-          }
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("postfixOp");
-        }
-        return result0;
-      }
-      
-      function parse_ValueOrParens() {
-        var result0, result1, result2, result3, result4;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        result0 = parse_Terminal();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, t) {
-              return helper.decorateTerminal(t, offset, t.literal || t.name);
-            })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          pos1 = clone(pos);
-          result0 = parse_ParensOpen();
-          if (result0 !== null) {
-            result1 = parse__();
-            if (result1 !== null) {
-              result2 = parse_NumericalExpression();
-              if (result2 !== null) {
-                result3 = parse__();
-                if (result3 !== null) {
-                  result4 = parse_ParensClose();
-                  if (result4 !== null) {
-                    result0 = [result0, result1, result2, result3, result4];
-                  } else {
-                    result0 = null;
-                    pos = clone(pos1);
-                  }
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, po, expression, pc) { 
-                var parens = new ast.ParensNode(expression);
-                parens.inBrackets = true;
-                expression._og_offsetFirst = expression.offsetFirst;
-                expression._og_offsetLast = expression.offsetLast;
-                parens._offsetFirst = offset;
-                parens._offsetLast = pc.offset;
-                parens.isMissingParensClose = (pc.offset >= inputLength);
-                return parens; 
-              })(pos0.offset, pos0.line, pos0.column, result0[0], result0[2], result0[4]);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_ParensOpen() {
-        var result0;
-        
-        reportFailures++;
-        if (input.charCodeAt(pos.offset) === 40) {
-          result0 = "(";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"(\"");
-          }
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("parensOpen");
-        }
-        return result0;
-      }
-      
-      function parse_ParensClose() {
-        var result0;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        if (input.charCodeAt(pos.offset) === 41) {
-          result0 = ")";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\")\"");
-          }
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column) {
-            return {offset: offset};
-          })(pos0.offset, pos0.line, pos0.column);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("parensClose");
-        }
-        return result0;
-      }
-      
-      function parse_Terminal() {
-        var result0;
-        
-        result0 = parse_Number();
-        if (result0 === null) {
-          result0 = parse_Identifier();
-        }
-        return result0;
-      }
-      
-      function parse_Identifier() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_IdentifierLiteral();
-        if (result0 !== null) {
-          result1 = (function(offset, line, column, identifier) { return parseContext.scope.findVariable(identifier); })(pos.offset, pos.line, pos.column, result0) ? "" : null;
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, identifier) { 
-            return new ast.IdentifierNode(identifier, parseContext.scope);
-          })(pos0.offset, pos0.line, pos0.column, result0[0]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("variable");
-        }
-        return result0;
-      }
-      
-      function parse_UnitIdentifierLiteral() {
-        var result0;
-        
-        result0 = parse_SpecialUnitLiteral();
-        if (result0 === null) {
-          result0 = parse_IdentifierLiteral();
-        }
-        return result0;
-      }
-      
-      function parse_SpecialUnitLiteral() {
-        var result0;
-        
-        if (input.charCodeAt(pos.offset) === 65020) {
-          result0 = "\uFDFC";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"\\uFDFC\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 8353) {
-            result0 = "\u20A1";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"\\u20A1\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.charCodeAt(pos.offset) === 8366) {
-              result0 = "\u20AE";
-              advance(pos, 1);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"\\u20AE\"");
-              }
-            }
-            if (result0 === null) {
-              if (input.charCodeAt(pos.offset) === 8358) {
-                result0 = "\u20A6";
-                advance(pos, 1);
-              } else {
-                result0 = null;
-                if (reportFailures === 0) {
-                  matchFailed("\"\\u20A6\"");
-                }
-              }
-              if (result0 === null) {
-                if (input.charCodeAt(pos.offset) === 36) {
-                  result0 = "$";
-                  advance(pos, 1);
-                } else {
-                  result0 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("\"$\"");
-                  }
-                }
-                if (result0 === null) {
-                  if (input.charCodeAt(pos.offset) === 163) {
-                    result0 = "\xA3";
-                    advance(pos, 1);
-                  } else {
-                    result0 = null;
-                    if (reportFailures === 0) {
-                      matchFailed("\"\\xA3\"");
-                    }
-                  }
-                  if (result0 === null) {
-                    if (input.charCodeAt(pos.offset) === 8377) {
-                      result0 = "\u20B9";
-                      advance(pos, 1);
-                    } else {
-                      result0 = null;
-                      if (reportFailures === 0) {
-                        matchFailed("\"\\u20B9\"");
-                      }
-                    }
-                    if (result0 === null) {
-                      if (input.charCodeAt(pos.offset) === 165) {
-                        result0 = "\xA5";
-                        advance(pos, 1);
-                      } else {
-                        result0 = null;
-                        if (reportFailures === 0) {
-                          matchFailed("\"\\xA5\"");
-                        }
-                      }
-                      if (result0 === null) {
-                        if (input.charCodeAt(pos.offset) === 8365) {
-                          result0 = "\u20AD";
-                          advance(pos, 1);
-                        } else {
-                          result0 = null;
-                          if (reportFailures === 0) {
-                            matchFailed("\"\\u20AD\"");
-                          }
-                        }
-                        if (result0 === null) {
-                          if (input.charCodeAt(pos.offset) === 402) {
-                            result0 = "\u0192";
-                            advance(pos, 1);
-                          } else {
-                            result0 = null;
-                            if (reportFailures === 0) {
-                              matchFailed("\"\\u0192\"");
-                            }
-                          }
-                          if (result0 === null) {
-                            if (input.charCodeAt(pos.offset) === 8361) {
-                              result0 = "\u20A9";
-                              advance(pos, 1);
-                            } else {
-                              result0 = null;
-                              if (reportFailures === 0) {
-                                matchFailed("\"\\u20A9\"");
-                              }
-                            }
-                            if (result0 === null) {
-                              if (input.charCodeAt(pos.offset) === 8360) {
-                                result0 = "\u20A8";
-                                advance(pos, 1);
-                              } else {
-                                result0 = null;
-                                if (reportFailures === 0) {
-                                  matchFailed("\"\\u20A8\"");
-                                }
-                              }
-                              if (result0 === null) {
-                                if (input.charCodeAt(pos.offset) === 8364) {
-                                  result0 = "\u20AC";
-                                  advance(pos, 1);
-                                } else {
-                                  result0 = null;
-                                  if (reportFailures === 0) {
-                                    matchFailed("\"\\u20AC\"");
-                                  }
-                                }
-                                if (result0 === null) {
-                                  if (input.charCodeAt(pos.offset) === 8362) {
-                                    result0 = "\u20AA";
-                                    advance(pos, 1);
-                                  } else {
-                                    result0 = null;
-                                    if (reportFailures === 0) {
-                                      matchFailed("\"\\u20AA\"");
-                                    }
-                                  }
-                                  if (result0 === null) {
-                                    if (input.charCodeAt(pos.offset) === 8369) {
-                                      result0 = "\u20B1";
-                                      advance(pos, 1);
-                                    } else {
-                                      result0 = null;
-                                      if (reportFailures === 0) {
-                                        matchFailed("\"\\u20B1\"");
-                                      }
-                                    }
-                                    if (result0 === null) {
-                                      if (input.charCodeAt(pos.offset) === 3647) {
-                                        result0 = "\u0E3F";
-                                        advance(pos, 1);
-                                      } else {
-                                        result0 = null;
-                                        if (reportFailures === 0) {
-                                          matchFailed("\"\\u0E3F\"");
-                                        }
-                                      }
-                                      if (result0 === null) {
-                                        if (input.charCodeAt(pos.offset) === 1547) {
-                                          result0 = "\u060B";
-                                          advance(pos, 1);
-                                        } else {
-                                          result0 = null;
-                                          if (reportFailures === 0) {
-                                            matchFailed("\"\\u060B\"");
-                                          }
-                                        }
-                                        if (result0 === null) {
-                                          if (input.charCodeAt(pos.offset) === 6107) {
-                                            result0 = "\u17DB";
-                                            advance(pos, 1);
-                                          } else {
-                                            result0 = null;
-                                            if (reportFailures === 0) {
-                                              matchFailed("\"\\u17DB\"");
-                                            }
-                                          }
-                                          if (result0 === null) {
-                                            if (input.charCodeAt(pos.offset) === 8363) {
-                                              result0 = "\u20AB";
-                                              advance(pos, 1);
-                                            } else {
-                                              result0 = null;
-                                              if (reportFailures === 0) {
-                                                matchFailed("\"\\u20AB\"");
-                                              }
-                                            }
-                                            if (result0 === null) {
-                                              if (input.charCodeAt(pos.offset) === 8372) {
-                                                result0 = "\u20B4";
-                                                advance(pos, 1);
-                                              } else {
-                                                result0 = null;
-                                                if (reportFailures === 0) {
-                                                  matchFailed("\"\\u20B4\"");
-                                                }
-                                              }
-                                            }
-                                          }
-                                        }
-                                      }
-                                    }
-                                  }
-                                }
-                              }
-                            }
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_IdentifierLiteral() {
-        var result0, result1, result2, result3;
-        var pos0, pos1, pos2, pos3;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        pos2 = clone(pos);
-        reportFailures++;
-        pos3 = clone(pos);
-        result0 = parse_KeywordLiteral();
-        if (result0 !== null) {
-          result1 = parse___();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos3);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos3);
-        }
-        reportFailures--;
-        if (result0 === null) {
-          result0 = "";
-        } else {
-          result0 = null;
-          pos = clone(pos2);
-        }
-        if (result0 !== null) {
-          result1 = parse_IdentifierStart();
-          if (result1 !== null) {
-            result2 = [];
-            result3 = parse_IdentifierPart();
-            while (result3 !== null) {
-              result2.push(result3);
-              result3 = parse_IdentifierPart();
-            }
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, head, tail) { return head + tail.join(""); })(pos0.offset, pos0.line, pos0.column, result0[1], result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("identifier");
-        }
-        return result0;
-      }
-      
-      function parse_Number() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_NumberLiteralString();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, number) { 
-            var node = new ast.NumberNode(number.literal);
-            node.exponent = number.exponent;
-            node.integerPart = number.integerPart;
-            node.sign = number.sign;
-            return node;
-          })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NumberLiteral() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_NumberLiteralString();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, numberObj) {
-          return parseFloat(numberObj.literal , 10);
-        })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NumberLiteralString() {
-        var result0, result1, result2;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_AdditiveOperatorLiteral();
-        result0 = result0 !== null ? result0 : "";
-        if (result0 !== null) {
-          result1 = parse_NumberLiteralFromParts();
-          if (result1 !== null) {
-            result2 = parse_ExponentLiteral();
-            result2 = result2 !== null ? result2 : "";
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, sign, valueNode, exp) {
-          valueNode.exponent = exp;
-          valueNode.number = valueNode.literal;
-          valueNode.sign = sign;
-          
-          var literal = sign + valueNode.literal; 
-          if (exp) {
-            literal += 'e' + exp;
-          }
-          valueNode.literal = literal;
-          return valueNode;
-        })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1], result0[2]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_NumberLiteralFromParts() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_NumberWholePart();
-        if (result0 !== null) {
-          result1 = parse_NumberDecimalPart();
-          result1 = result1 !== null ? result1 : "";
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, whole, decimal) {
-              return {
-                mantissa: decimal,
-                integerPart: whole,
-                literal: whole + decimal
-              }
-            })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        if (result0 === null) {
-          pos0 = clone(pos);
-          result0 = parse_NumberDecimalPart();
-          if (result0 !== null) {
-            result0 = (function(offset, line, column, decimal) {
-                return {
-                  mantissa: decimal,
-                  integerPart: '0',
-                  literal: decimal
-                }
-              })(pos0.offset, pos0.line, pos0.column, result0);
-          }
-          if (result0 === null) {
-            pos = clone(pos0);
-          }
-        }
-        return result0;
-      }
-      
-      function parse_NumberWholePart() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_Digits();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, digits) {
-          return digits;
-        })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_Digits() {
-        var result0, result1;
-        var pos0;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        if (/^[0-9]/.test(input.charAt(pos.offset))) {
-          result1 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result1 = null;
-          if (reportFailures === 0) {
-            matchFailed("[0-9]");
-          }
-        }
-        if (result1 !== null) {
-          result0 = [];
-          while (result1 !== null) {
-            result0.push(result1);
-            if (/^[0-9]/.test(input.charAt(pos.offset))) {
-              result1 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result1 = null;
-              if (reportFailures === 0) {
-                matchFailed("[0-9]");
-              }
-            }
-          }
-        } else {
-          result0 = null;
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, digits) {
-          // this is helpful for knowing if we have EOF e.g. keyboard, autocomplete
-          self.lastDigitOffset = offset + digits.length;
-          return digits.join("");
-        })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("digits");
-        }
-        return result0;
-      }
-      
-      function parse_NumberDecimalPart() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_DecimalPoint();
-        if (result0 !== null) {
-          result1 = parse_Digits();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, digits) {
-          return "." + digits;
-        })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_DecimalPoint() {
-        var result0;
-        
-        reportFailures++;
-        if (input.charCodeAt(pos.offset) === 46) {
-          result0 = ".";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\".\"");
-          }
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("point");
-        }
-        return result0;
-      }
-      
-      function parse_ExponentLiteral() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_ExponentSign();
-        if (result0 !== null) {
-          result1 = parse_IntegerString();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, integer) {
-          return integer;
-        })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_ExponentSign() {
-        var result0;
-        
-        reportFailures++;
-        if (input.charCodeAt(pos.offset) === 101) {
-          result0 = "e";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"e\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 69) {
-            result0 = "E";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"E\"");
-            }
-          }
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("exponent");
-        }
-        return result0;
-      }
-      
-      function parse_IntegerNode() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_IntegerString();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, number) { 
-          var node = new ast.NumberNode(number);
-          return helper.decorateTerminal(node, offset, number);
-        })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_IntegerLiteral() {
-        var result0;
-        var pos0;
-        
-        pos0 = clone(pos);
-        result0 = parse_IntegerString();
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, string) {
-          return parseInt(string, 10);
-        })(pos0.offset, pos0.line, pos0.column, result0);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_IntegerString() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        result0 = parse_AdditiveOperatorLiteral();
-        result0 = result0 !== null ? result0 : "";
-        if (result0 !== null) {
-          result1 = parse_NumberWholePart();
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, sign, digits) {
-          return sign + digits;
-        })(pos0.offset, pos0.line, pos0.column, result0[0], result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_KeywordLiteral() {
-        var result0;
-        
-        if (input.substr(pos.offset, 4) === "unit") {
-          result0 = "unit";
-          advance(pos, 4);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"unit\"");
-          }
-        }
-        if (result0 === null) {
-          if (input.substr(pos.offset, 3) === "per") {
-            result0 = "per";
-            advance(pos, 3);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"per\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.substr(pos.offset, 5) === "const") {
-              result0 = "const";
-              advance(pos, 5);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"const\"");
-              }
-            }
-            if (result0 === null) {
-              if (input.substr(pos.offset, 4) === "test") {
-                result0 = "test";
-                advance(pos, 4);
-              } else {
-                result0 = null;
-                if (reportFailures === 0) {
-                  matchFailed("\"test\"");
-                }
-              }
-              if (result0 === null) {
-                if (input.substr(pos.offset, 7) === "include") {
-                  result0 = "include";
-                  advance(pos, 7);
-                } else {
-                  result0 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("\"include\"");
-                  }
-                }
-                if (result0 === null) {
-                  if (input.substr(pos.offset, 5) === "units") {
-                    result0 = "units";
-                    advance(pos, 5);
-                  } else {
-                    result0 = null;
-                    if (reportFailures === 0) {
-                      matchFailed("\"units\"");
-                    }
-                  }
-                  if (result0 === null) {
-                    if (input.substr(pos.offset, 3) === "let") {
-                      result0 = "let";
-                      advance(pos, 3);
-                    } else {
-                      result0 = null;
-                      if (reportFailures === 0) {
-                        matchFailed("\"let\"");
-                      }
-                    }
-                    if (result0 === null) {
-                      if (input.substr(pos.offset, 2) === "in") {
-                        result0 = "in";
-                        advance(pos, 2);
-                      } else {
-                        result0 = null;
-                        if (reportFailures === 0) {
-                          matchFailed("\"in\"");
-                        }
-                      }
-                      if (result0 === null) {
-                        if (input.substr(pos.offset, 2) === "at") {
-                          result0 = "at";
-                          advance(pos, 2);
-                        } else {
-                          result0 = null;
-                          if (reportFailures === 0) {
-                            matchFailed("\"at\"");
-                          }
-                        }
-                        if (result0 === null) {
-                          if (input.substr(pos.offset, 2) === "of") {
-                            result0 = "of";
-                            advance(pos, 2);
-                          } else {
-                            result0 = null;
-                            if (reportFailures === 0) {
-                              matchFailed("\"of\"");
-                            }
-                          }
-                          if (result0 === null) {
-                            if (input.substr(pos.offset, 5) === "every") {
-                              result0 = "every";
-                              advance(pos, 5);
-                            } else {
-                              result0 = null;
-                              if (reportFailures === 0) {
-                                matchFailed("\"every\"");
-                              }
-                            }
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse__() {
-        var result0, result1, result2, result3, result4;
-        var pos0, pos1;
-        
-        reportFailures++;
-        pos0 = clone(pos);
-        result0 = [];
-        if (/^[ \t\r\n\xA0]/.test(input.charAt(pos.offset))) {
-          result1 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result1 = null;
-          if (reportFailures === 0) {
-            matchFailed("[ \\t\\r\\n\\xA0]");
-          }
-        }
-        while (result1 !== null) {
-          result0.push(result1);
-          if (/^[ \t\r\n\xA0]/.test(input.charAt(pos.offset))) {
-            result1 = input.charAt(pos.offset);
-            advance(pos, 1);
-          } else {
-            result1 = null;
-            if (reportFailures === 0) {
-              matchFailed("[ \\t\\r\\n\\xA0]");
-            }
-          }
-        }
-        if (result0 !== null) {
-          result1 = [];
-          pos1 = clone(pos);
-          result2 = parse_Comment();
-          if (result2 !== null) {
-            result3 = [];
-            if (/^[ \t\r\n]/.test(input.charAt(pos.offset))) {
-              result4 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result4 = null;
-              if (reportFailures === 0) {
-                matchFailed("[ \\t\\r\\n]");
-              }
-            }
-            while (result4 !== null) {
-              result3.push(result4);
-              if (/^[ \t\r\n]/.test(input.charAt(pos.offset))) {
-                result4 = input.charAt(pos.offset);
-                advance(pos, 1);
-              } else {
-                result4 = null;
-                if (reportFailures === 0) {
-                  matchFailed("[ \\t\\r\\n]");
-                }
-              }
-            }
-            if (result3 !== null) {
-              result2 = [result2, result3];
-            } else {
-              result2 = null;
-              pos = clone(pos1);
-            }
-          } else {
-            result2 = null;
-            pos = clone(pos1);
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            pos1 = clone(pos);
-            result2 = parse_Comment();
-            if (result2 !== null) {
-              result3 = [];
-              if (/^[ \t\r\n]/.test(input.charAt(pos.offset))) {
-                result4 = input.charAt(pos.offset);
-                advance(pos, 1);
-              } else {
-                result4 = null;
-                if (reportFailures === 0) {
-                  matchFailed("[ \\t\\r\\n]");
-                }
-              }
-              while (result4 !== null) {
-                result3.push(result4);
-                if (/^[ \t\r\n]/.test(input.charAt(pos.offset))) {
-                  result4 = input.charAt(pos.offset);
-                  advance(pos, 1);
-                } else {
-                  result4 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("[ \\t\\r\\n]");
-                  }
-                }
-              }
-              if (result3 !== null) {
-                result2 = [result2, result3];
-              } else {
-                result2 = null;
-                pos = clone(pos1);
-              }
-            } else {
-              result2 = null;
-              pos = clone(pos1);
-            }
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos0);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos0);
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("some whitespace");
-        }
-        return result0;
-      }
-      
-      function parse_Comment() {
-        var result0, result1, result2, result3;
-        var pos0;
-        
-        pos0 = clone(pos);
-        if (input.substr(pos.offset, 2) === "//") {
-          result0 = "//";
-          advance(pos, 2);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"//\"");
-          }
-        }
-        if (result0 !== null) {
-          result1 = [];
-          if (/^[^\n]/.test(input.charAt(pos.offset))) {
-            result2 = input.charAt(pos.offset);
-            advance(pos, 1);
-          } else {
-            result2 = null;
-            if (reportFailures === 0) {
-              matchFailed("[^\\n]");
-            }
-          }
-          while (result2 !== null) {
-            result1.push(result2);
-            if (/^[^\n]/.test(input.charAt(pos.offset))) {
-              result2 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result2 = null;
-              if (reportFailures === 0) {
-                matchFailed("[^\\n]");
-              }
-            }
-          }
-          if (result1 !== null) {
-            if (/^[\r\n]/.test(input.charAt(pos.offset))) {
-              result3 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result3 = null;
-              if (reportFailures === 0) {
-                matchFailed("[\\r\\n]");
-              }
-            }
-            if (result3 !== null) {
-              result2 = [];
-              while (result3 !== null) {
-                result2.push(result3);
-                if (/^[\r\n]/.test(input.charAt(pos.offset))) {
-                  result3 = input.charAt(pos.offset);
-                  advance(pos, 1);
-                } else {
-                  result3 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("[\\r\\n]");
-                  }
-                }
-              }
-            } else {
-              result2 = null;
-            }
-            if (result2 !== null) {
-              result0 = [result0, result1, result2];
-            } else {
-              result0 = null;
-              pos = clone(pos0);
-            }
-          } else {
-            result0 = null;
-            pos = clone(pos0);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse___() {
-        var result0, result1;
-        
-        reportFailures++;
-        if (/^[ \t\r\n]/.test(input.charAt(pos.offset))) {
-          result1 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result1 = null;
-          if (reportFailures === 0) {
-            matchFailed("[ \\t\\r\\n]");
-          }
-        }
-        if (result1 !== null) {
-          result0 = [];
-          while (result1 !== null) {
-            result0.push(result1);
-            if (/^[ \t\r\n]/.test(input.charAt(pos.offset))) {
-              result1 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result1 = null;
-              if (reportFailures === 0) {
-                matchFailed("[ \\t\\r\\n]");
-              }
-            }
-          }
-        } else {
-          result0 = null;
-        }
-        reportFailures--;
-        if (reportFailures === 0 && result0 === null) {
-          matchFailed("whitespace");
-        }
-        return result0;
-      }
-      
-      function parse_nbsp() {
-        var result0, result1;
-        
-        result0 = [];
-        if (/^[ \t]/.test(input.charAt(pos.offset))) {
-          result1 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result1 = null;
-          if (reportFailures === 0) {
-            matchFailed("[ \\t]");
-          }
-        }
-        while (result1 !== null) {
-          result0.push(result1);
-          if (/^[ \t]/.test(input.charAt(pos.offset))) {
-            result1 = input.charAt(pos.offset);
-            advance(pos, 1);
-          } else {
-            result1 = null;
-            if (reportFailures === 0) {
-              matchFailed("[ \\t]");
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_nbsp_1() {
-        var result0, result1;
-        
-        if (/^[ \t]/.test(input.charAt(pos.offset))) {
-          result1 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result1 = null;
-          if (reportFailures === 0) {
-            matchFailed("[ \\t]");
-          }
-        }
-        if (result1 !== null) {
-          result0 = [];
-          while (result1 !== null) {
-            result0.push(result1);
-            if (/^[ \t]/.test(input.charAt(pos.offset))) {
-              result1 = input.charAt(pos.offset);
-              advance(pos, 1);
-            } else {
-              result1 = null;
-              if (reportFailures === 0) {
-                matchFailed("[ \\t]");
-              }
-            }
-          }
-        } else {
-          result0 = null;
-        }
-        return result0;
-      }
-      
-      function parse_IdentifierStart() {
-        var result0, result1;
-        var pos0, pos1;
-        
-        result0 = parse_UnicodeLetter();
-        if (result0 === null) {
-          if (input.charCodeAt(pos.offset) === 36) {
-            result0 = "$";
-            advance(pos, 1);
-          } else {
-            result0 = null;
-            if (reportFailures === 0) {
-              matchFailed("\"$\"");
-            }
-          }
-          if (result0 === null) {
-            if (input.charCodeAt(pos.offset) === 95) {
-              result0 = "_";
-              advance(pos, 1);
-            } else {
-              result0 = null;
-              if (reportFailures === 0) {
-                matchFailed("\"_\"");
-              }
-            }
-            if (result0 === null) {
-              pos0 = clone(pos);
-              pos1 = clone(pos);
-              if (input.charCodeAt(pos.offset) === 92) {
-                result0 = "\\";
-                advance(pos, 1);
-              } else {
-                result0 = null;
-                if (reportFailures === 0) {
-                  matchFailed("\"\\\\\"");
-                }
-              }
-              if (result0 !== null) {
-                result1 = parse_UnicodeEscapeSequence();
-                if (result1 !== null) {
-                  result0 = [result0, result1];
-                } else {
-                  result0 = null;
-                  pos = clone(pos1);
-                }
-              } else {
-                result0 = null;
-                pos = clone(pos1);
-              }
-              if (result0 !== null) {
-                result0 = (function(offset, line, column, sequence) { return sequence; })(pos0.offset, pos0.line, pos0.column, result0[1]);
-              }
-              if (result0 === null) {
-                pos = clone(pos0);
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_IdentifierPart() {
-        var result0;
-        
-        result0 = parse_IdentifierStart();
-        if (result0 === null) {
-          result0 = parse_UnicodeCombiningMark();
-          if (result0 === null) {
-            result0 = parse_Nd();
-            if (result0 === null) {
-              result0 = parse_Pc();
-              if (result0 === null) {
-                if (input.charCodeAt(pos.offset) === 8204) {
-                  result0 = "\u200C";
-                  advance(pos, 1);
-                } else {
-                  result0 = null;
-                  if (reportFailures === 0) {
-                    matchFailed("\"\\u200C\"");
-                  }
-                }
-                if (result0 === null) {
-                  if (input.charCodeAt(pos.offset) === 8205) {
-                    result0 = "\u200D";
-                    advance(pos, 1);
-                  } else {
-                    result0 = null;
-                    if (reportFailures === 0) {
-                      matchFailed("\"\\u200D\"");
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnicodeLetter() {
-        var result0;
-        
-        result0 = parse_Lu();
-        if (result0 === null) {
-          result0 = parse_Ll();
-          if (result0 === null) {
-            result0 = parse_Lt();
-            if (result0 === null) {
-              result0 = parse_Lm();
-              if (result0 === null) {
-                result0 = parse_Lo();
-                if (result0 === null) {
-                  result0 = parse_Nl();
-                }
-              }
-            }
-          }
-        }
-        return result0;
-      }
-      
-      function parse_UnicodeCombiningMark() {
-        var result0;
-        
-        result0 = parse_Mn();
-        if (result0 === null) {
-          result0 = parse_Mc();
-        }
-        return result0;
-      }
-      
-      function parse_UnicodeEscapeSequence() {
-        var result0, result1, result2, result3, result4;
-        var pos0, pos1, pos2;
-        
-        pos0 = clone(pos);
-        pos1 = clone(pos);
-        if (input.charCodeAt(pos.offset) === 117) {
-          result0 = "u";
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("\"u\"");
-          }
-        }
-        if (result0 !== null) {
-          pos2 = clone(pos);
-          result1 = parse_HexDigit();
-          if (result1 !== null) {
-            result2 = parse_HexDigit();
-            if (result2 !== null) {
-              result3 = parse_HexDigit();
-              if (result3 !== null) {
-                result4 = parse_HexDigit();
-                if (result4 !== null) {
-                  result1 = [result1, result2, result3, result4];
-                } else {
-                  result1 = null;
-                  pos = clone(pos2);
-                }
-              } else {
-                result1 = null;
-                pos = clone(pos2);
-              }
-            } else {
-              result1 = null;
-              pos = clone(pos2);
-            }
-          } else {
-            result1 = null;
-            pos = clone(pos2);
-          }
-          if (result1 !== null) {
-            result0 = [result0, result1];
-          } else {
-            result0 = null;
-            pos = clone(pos1);
-          }
-        } else {
-          result0 = null;
-          pos = clone(pos1);
-        }
-        if (result0 !== null) {
-          result0 = (function(offset, line, column, digits) {
-              digits = digits.join('');
-              return String.fromCharCode(parseInt(digits, 16));
-            })(pos0.offset, pos0.line, pos0.column, result0[1]);
-        }
-        if (result0 === null) {
-          pos = clone(pos0);
-        }
-        return result0;
-      }
-      
-      function parse_HexDigit() {
-        var result0;
-        
-        if (/^[0-9a-f]/i.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[0-9a-f]i");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Ll() {
-        var result0;
-        
-        if (/^[a-z\xB5\xDF-\xF6\xF8-\xFF\u0101\u0103\u0105\u0107\u0109\u010B\u010D\u010F\u0111\u0113\u0115\u0117\u0119\u011B\u011D\u011F\u0121\u0123\u0125\u0127\u0129\u012B\u012D\u012F\u0131\u0133\u0135\u0137-\u0138\u013A\u013C\u013E\u0140\u0142\u0144\u0146\u0148-\u0149\u014B\u014D\u014F\u0151\u0153\u0155\u0157\u0159\u015B\u015D\u015F\u0161\u0163\u0165\u0167\u0169\u016B\u016D\u016F\u0171\u0173\u0175\u0177\u017A\u017C\u017E-\u0180\u0183\u0185\u0188\u018C-\u018D\u0192\u0195\u0199-\u019B\u019E\u01A1\u01A3\u01A5\u01A8\u01AA-\u01AB\u01AD\u01B0\u01B4\u01B6\u01B9-\u01BA\u01BD-\u01BF\u01C6\u01C9\u01CC\u01CE\u01D0\u01D2\u01D4\u01D6\u01D8\u01DA\u01DC-\u01DD\u01DF\u01E1\u01E3\u01E5\u01E7\u01E9\u01EB\u01ED\u01EF-\u01F0\u01F3\u01F5\u01F9\u01FB\u01FD\u01FF\u0201\u0203\u0205\u0207\u0209\u020B\u020D\u020F\u0211\u0213\u0215\u0217\u0219\u021B\u021D\u021F\u0221\u0223\u0225\u0227\u0229\u022B\u022D\u022F\u0231\u0233-\u0239\u023C\u023F-\u0240\u0242\u0247\u0249\u024B\u024D\u024F-\u0293\u0295-\u02AF\u0371\u0373\u0377\u037B-\u037D\u0390\u03AC-\u03CE\u03D0-\u03D1\u03D5-\u03D7\u03D9\u03DB\u03DD\u03DF\u03E1\u03E3\u03E5\u03E7\u03E9\u03EB\u03ED\u03EF-\u03F3\u03F5\u03F8\u03FB-\u03FC\u0430-\u045F\u0461\u0463\u0465\u0467\u0469\u046B\u046D\u046F\u0471\u0473\u0475\u0477\u0479\u047B\u047D\u047F\u0481\u048B\u048D\u048F\u0491\u0493\u0495\u0497\u0499\u049B\u049D\u049F\u04A1\u04A3\u04A5\u04A7\u04A9\u04AB\u04AD\u04AF\u04B1\u04B3\u04B5\u04B7\u04B9\u04BB\u04BD\u04BF\u04C2\u04C4\u04C6\u04C8\u04CA\u04CC\u04CE-\u04CF\u04D1\u04D3\u04D5\u04D7\u04D9\u04DB\u04DD\u04DF\u04E1\u04E3\u04E5\u04E7\u04E9\u04EB\u04ED\u04EF\u04F1\u04F3\u04F5\u04F7\u04F9\u04FB\u04FD\u04FF\u0501\u0503\u0505\u0507\u0509\u050B\u050D\u050F\u0511\u0513\u0515\u0517\u0519\u051B\u051D\u051F\u0521\u0523\u0525\u0527\u0561-\u0587\u1D00-\u1D2B\u1D6B-\u1D77\u1D79-\u1D9A\u1E01\u1E03\u1E05\u1E07\u1E09\u1E0B\u1E0D\u1E0F\u1E11\u1E13\u1E15\u1E17\u1E19\u1E1B\u1E1D\u1E1F\u1E21\u1E23\u1E25\u1E27\u1E29\u1E2B\u1E2D\u1E2F\u1E31\u1E33\u1E35\u1E37\u1E39\u1E3B\u1E3D\u1E3F\u1E41\u1E43\u1E45\u1E47\u1E49\u1E4B\u1E4D\u1E4F\u1E51\u1E53\u1E55\u1E57\u1E59\u1E5B\u1E5D\u1E5F\u1E61\u1E63\u1E65\u1E67\u1E69\u1E6B\u1E6D\u1E6F\u1E71\u1E73\u1E75\u1E77\u1E79\u1E7B\u1E7D\u1E7F\u1E81\u1E83\u1E85\u1E87\u1E89\u1E8B\u1E8D\u1E8F\u1E91\u1E93\u1E95-\u1E9D\u1E9F\u1EA1\u1EA3\u1EA5\u1EA7\u1EA9\u1EAB\u1EAD\u1EAF\u1EB1\u1EB3\u1EB5\u1EB7\u1EB9\u1EBB\u1EBD\u1EBF\u1EC1\u1EC3\u1EC5\u1EC7\u1EC9\u1ECB\u1ECD\u1ECF\u1ED1\u1ED3\u1ED5\u1ED7\u1ED9\u1EDB\u1EDD\u1EDF\u1EE1\u1EE3\u1EE5\u1EE7\u1EE9\u1EEB\u1EED\u1EEF\u1EF1\u1EF3\u1EF5\u1EF7\u1EF9\u1EFB\u1EFD\u1EFF-\u1F07\u1F10-\u1F15\u1F20-\u1F27\u1F30-\u1F37\u1F40-\u1F45\u1F50-\u1F57\u1F60-\u1F67\u1F70-\u1F7D\u1F80-\u1F87\u1F90-\u1F97\u1FA0-\u1FA7\u1FB0-\u1FB4\u1FB6-\u1FB7\u1FBE\u1FC2-\u1FC4\u1FC6-\u1FC7\u1FD0-\u1FD3\u1FD6-\u1FD7\u1FE0-\u1FE7\u1FF2-\u1FF4\u1FF6-\u1FF7\u210A\u210E-\u210F\u2113\u212F\u2134\u2139\u213C-\u213D\u2146-\u2149\u214E\u2184\u2C30-\u2C5E\u2C61\u2C65-\u2C66\u2C68\u2C6A\u2C6C\u2C71\u2C73-\u2C74\u2C76-\u2C7B\u2C81\u2C83\u2C85\u2C87\u2C89\u2C8B\u2C8D\u2C8F\u2C91\u2C93\u2C95\u2C97\u2C99\u2C9B\u2C9D\u2C9F\u2CA1\u2CA3\u2CA5\u2CA7\u2CA9\u2CAB\u2CAD\u2CAF\u2CB1\u2CB3\u2CB5\u2CB7\u2CB9\u2CBB\u2CBD\u2CBF\u2CC1\u2CC3\u2CC5\u2CC7\u2CC9\u2CCB\u2CCD\u2CCF\u2CD1\u2CD3\u2CD5\u2CD7\u2CD9\u2CDB\u2CDD\u2CDF\u2CE1\u2CE3-\u2CE4\u2CEC\u2CEE\u2CF3\u2D00-\u2D25\u2D27\u2D2D\uA641\uA643\uA645\uA647\uA649\uA64B\uA64D\uA64F\uA651\uA653\uA655\uA657\uA659\uA65B\uA65D\uA65F\uA661\uA663\uA665\uA667\uA669\uA66B\uA66D\uA681\uA683\uA685\uA687\uA689\uA68B\uA68D\uA68F\uA691\uA693\uA695\uA697\uA723\uA725\uA727\uA729\uA72B\uA72D\uA72F-\uA731\uA733\uA735\uA737\uA739\uA73B\uA73D\uA73F\uA741\uA743\uA745\uA747\uA749\uA74B\uA74D\uA74F\uA751\uA753\uA755\uA757\uA759\uA75B\uA75D\uA75F\uA761\uA763\uA765\uA767\uA769\uA76B\uA76D\uA76F\uA771-\uA778\uA77A\uA77C\uA77F\uA781\uA783\uA785\uA787\uA78C\uA78E\uA791\uA793\uA7A1\uA7A3\uA7A5\uA7A7\uA7A9\uA7FA\uFB00-\uFB06\uFB13-\uFB17\uFF41-\uFF5A]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[a-z\\xB5\\xDF-\\xF6\\xF8-\\xFF\\u0101\\u0103\\u0105\\u0107\\u0109\\u010B\\u010D\\u010F\\u0111\\u0113\\u0115\\u0117\\u0119\\u011B\\u011D\\u011F\\u0121\\u0123\\u0125\\u0127\\u0129\\u012B\\u012D\\u012F\\u0131\\u0133\\u0135\\u0137-\\u0138\\u013A\\u013C\\u013E\\u0140\\u0142\\u0144\\u0146\\u0148-\\u0149\\u014B\\u014D\\u014F\\u0151\\u0153\\u0155\\u0157\\u0159\\u015B\\u015D\\u015F\\u0161\\u0163\\u0165\\u0167\\u0169\\u016B\\u016D\\u016F\\u0171\\u0173\\u0175\\u0177\\u017A\\u017C\\u017E-\\u0180\\u0183\\u0185\\u0188\\u018C-\\u018D\\u0192\\u0195\\u0199-\\u019B\\u019E\\u01A1\\u01A3\\u01A5\\u01A8\\u01AA-\\u01AB\\u01AD\\u01B0\\u01B4\\u01B6\\u01B9-\\u01BA\\u01BD-\\u01BF\\u01C6\\u01C9\\u01CC\\u01CE\\u01D0\\u01D2\\u01D4\\u01D6\\u01D8\\u01DA\\u01DC-\\u01DD\\u01DF\\u01E1\\u01E3\\u01E5\\u01E7\\u01E9\\u01EB\\u01ED\\u01EF-\\u01F0\\u01F3\\u01F5\\u01F9\\u01FB\\u01FD\\u01FF\\u0201\\u0203\\u0205\\u0207\\u0209\\u020B\\u020D\\u020F\\u0211\\u0213\\u0215\\u0217\\u0219\\u021B\\u021D\\u021F\\u0221\\u0223\\u0225\\u0227\\u0229\\u022B\\u022D\\u022F\\u0231\\u0233-\\u0239\\u023C\\u023F-\\u0240\\u0242\\u0247\\u0249\\u024B\\u024D\\u024F-\\u0293\\u0295-\\u02AF\\u0371\\u0373\\u0377\\u037B-\\u037D\\u0390\\u03AC-\\u03CE\\u03D0-\\u03D1\\u03D5-\\u03D7\\u03D9\\u03DB\\u03DD\\u03DF\\u03E1\\u03E3\\u03E5\\u03E7\\u03E9\\u03EB\\u03ED\\u03EF-\\u03F3\\u03F5\\u03F8\\u03FB-\\u03FC\\u0430-\\u045F\\u0461\\u0463\\u0465\\u0467\\u0469\\u046B\\u046D\\u046F\\u0471\\u0473\\u0475\\u0477\\u0479\\u047B\\u047D\\u047F\\u0481\\u048B\\u048D\\u048F\\u0491\\u0493\\u0495\\u0497\\u0499\\u049B\\u049D\\u049F\\u04A1\\u04A3\\u04A5\\u04A7\\u04A9\\u04AB\\u04AD\\u04AF\\u04B1\\u04B3\\u04B5\\u04B7\\u04B9\\u04BB\\u04BD\\u04BF\\u04C2\\u04C4\\u04C6\\u04C8\\u04CA\\u04CC\\u04CE-\\u04CF\\u04D1\\u04D3\\u04D5\\u04D7\\u04D9\\u04DB\\u04DD\\u04DF\\u04E1\\u04E3\\u04E5\\u04E7\\u04E9\\u04EB\\u04ED\\u04EF\\u04F1\\u04F3\\u04F5\\u04F7\\u04F9\\u04FB\\u04FD\\u04FF\\u0501\\u0503\\u0505\\u0507\\u0509\\u050B\\u050D\\u050F\\u0511\\u0513\\u0515\\u0517\\u0519\\u051B\\u051D\\u051F\\u0521\\u0523\\u0525\\u0527\\u0561-\\u0587\\u1D00-\\u1D2B\\u1D6B-\\u1D77\\u1D79-\\u1D9A\\u1E01\\u1E03\\u1E05\\u1E07\\u1E09\\u1E0B\\u1E0D\\u1E0F\\u1E11\\u1E13\\u1E15\\u1E17\\u1E19\\u1E1B\\u1E1D\\u1E1F\\u1E21\\u1E23\\u1E25\\u1E27\\u1E29\\u1E2B\\u1E2D\\u1E2F\\u1E31\\u1E33\\u1E35\\u1E37\\u1E39\\u1E3B\\u1E3D\\u1E3F\\u1E41\\u1E43\\u1E45\\u1E47\\u1E49\\u1E4B\\u1E4D\\u1E4F\\u1E51\\u1E53\\u1E55\\u1E57\\u1E59\\u1E5B\\u1E5D\\u1E5F\\u1E61\\u1E63\\u1E65\\u1E67\\u1E69\\u1E6B\\u1E6D\\u1E6F\\u1E71\\u1E73\\u1E75\\u1E77\\u1E79\\u1E7B\\u1E7D\\u1E7F\\u1E81\\u1E83\\u1E85\\u1E87\\u1E89\\u1E8B\\u1E8D\\u1E8F\\u1E91\\u1E93\\u1E95-\\u1E9D\\u1E9F\\u1EA1\\u1EA3\\u1EA5\\u1EA7\\u1EA9\\u1EAB\\u1EAD\\u1EAF\\u1EB1\\u1EB3\\u1EB5\\u1EB7\\u1EB9\\u1EBB\\u1EBD\\u1EBF\\u1EC1\\u1EC3\\u1EC5\\u1EC7\\u1EC9\\u1ECB\\u1ECD\\u1ECF\\u1ED1\\u1ED3\\u1ED5\\u1ED7\\u1ED9\\u1EDB\\u1EDD\\u1EDF\\u1EE1\\u1EE3\\u1EE5\\u1EE7\\u1EE9\\u1EEB\\u1EED\\u1EEF\\u1EF1\\u1EF3\\u1EF5\\u1EF7\\u1EF9\\u1EFB\\u1EFD\\u1EFF-\\u1F07\\u1F10-\\u1F15\\u1F20-\\u1F27\\u1F30-\\u1F37\\u1F40-\\u1F45\\u1F50-\\u1F57\\u1F60-\\u1F67\\u1F70-\\u1F7D\\u1F80-\\u1F87\\u1F90-\\u1F97\\u1FA0-\\u1FA7\\u1FB0-\\u1FB4\\u1FB6-\\u1FB7\\u1FBE\\u1FC2-\\u1FC4\\u1FC6-\\u1FC7\\u1FD0-\\u1FD3\\u1FD6-\\u1FD7\\u1FE0-\\u1FE7\\u1FF2-\\u1FF4\\u1FF6-\\u1FF7\\u210A\\u210E-\\u210F\\u2113\\u212F\\u2134\\u2139\\u213C-\\u213D\\u2146-\\u2149\\u214E\\u2184\\u2C30-\\u2C5E\\u2C61\\u2C65-\\u2C66\\u2C68\\u2C6A\\u2C6C\\u2C71\\u2C73-\\u2C74\\u2C76-\\u2C7B\\u2C81\\u2C83\\u2C85\\u2C87\\u2C89\\u2C8B\\u2C8D\\u2C8F\\u2C91\\u2C93\\u2C95\\u2C97\\u2C99\\u2C9B\\u2C9D\\u2C9F\\u2CA1\\u2CA3\\u2CA5\\u2CA7\\u2CA9\\u2CAB\\u2CAD\\u2CAF\\u2CB1\\u2CB3\\u2CB5\\u2CB7\\u2CB9\\u2CBB\\u2CBD\\u2CBF\\u2CC1\\u2CC3\\u2CC5\\u2CC7\\u2CC9\\u2CCB\\u2CCD\\u2CCF\\u2CD1\\u2CD3\\u2CD5\\u2CD7\\u2CD9\\u2CDB\\u2CDD\\u2CDF\\u2CE1\\u2CE3-\\u2CE4\\u2CEC\\u2CEE\\u2CF3\\u2D00-\\u2D25\\u2D27\\u2D2D\\uA641\\uA643\\uA645\\uA647\\uA649\\uA64B\\uA64D\\uA64F\\uA651\\uA653\\uA655\\uA657\\uA659\\uA65B\\uA65D\\uA65F\\uA661\\uA663\\uA665\\uA667\\uA669\\uA66B\\uA66D\\uA681\\uA683\\uA685\\uA687\\uA689\\uA68B\\uA68D\\uA68F\\uA691\\uA693\\uA695\\uA697\\uA723\\uA725\\uA727\\uA729\\uA72B\\uA72D\\uA72F-\\uA731\\uA733\\uA735\\uA737\\uA739\\uA73B\\uA73D\\uA73F\\uA741\\uA743\\uA745\\uA747\\uA749\\uA74B\\uA74D\\uA74F\\uA751\\uA753\\uA755\\uA757\\uA759\\uA75B\\uA75D\\uA75F\\uA761\\uA763\\uA765\\uA767\\uA769\\uA76B\\uA76D\\uA76F\\uA771-\\uA778\\uA77A\\uA77C\\uA77F\\uA781\\uA783\\uA785\\uA787\\uA78C\\uA78E\\uA791\\uA793\\uA7A1\\uA7A3\\uA7A5\\uA7A7\\uA7A9\\uA7FA\\uFB00-\\uFB06\\uFB13-\\uFB17\\uFF41-\\uFF5A]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Lm() {
-        var result0;
-        
-        if (/^[\u02B0-\u02C1\u02C6-\u02D1\u02E0-\u02E4\u02EC\u02EE\u0374\u037A\u0559\u0640\u06E5-\u06E6\u07F4-\u07F5\u07FA\u081A\u0824\u0828\u0971\u0E46\u0EC6\u10FC\u17D7\u1843\u1AA7\u1C78-\u1C7D\u1D2C-\u1D6A\u1D78\u1D9B-\u1DBF\u2071\u207F\u2090-\u209C\u2C7C-\u2C7D\u2D6F\u2E2F\u3005\u3031-\u3035\u303B\u309D-\u309E\u30FC-\u30FE\uA015\uA4F8-\uA4FD\uA60C\uA67F\uA717-\uA71F\uA770\uA788\uA7F8-\uA7F9\uA9CF\uAA70\uAADD\uAAF3-\uAAF4\uFF70\uFF9E-\uFF9F]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\u02B0-\\u02C1\\u02C6-\\u02D1\\u02E0-\\u02E4\\u02EC\\u02EE\\u0374\\u037A\\u0559\\u0640\\u06E5-\\u06E6\\u07F4-\\u07F5\\u07FA\\u081A\\u0824\\u0828\\u0971\\u0E46\\u0EC6\\u10FC\\u17D7\\u1843\\u1AA7\\u1C78-\\u1C7D\\u1D2C-\\u1D6A\\u1D78\\u1D9B-\\u1DBF\\u2071\\u207F\\u2090-\\u209C\\u2C7C-\\u2C7D\\u2D6F\\u2E2F\\u3005\\u3031-\\u3035\\u303B\\u309D-\\u309E\\u30FC-\\u30FE\\uA015\\uA4F8-\\uA4FD\\uA60C\\uA67F\\uA717-\\uA71F\\uA770\\uA788\\uA7F8-\\uA7F9\\uA9CF\\uAA70\\uAADD\\uAAF3-\\uAAF4\\uFF70\\uFF9E-\\uFF9F]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Lo() {
-        var result0;
-        
-        if (/^[\xAA\xBA\u01BB\u01C0-\u01C3\u0294\u05D0-\u05EA\u05F0-\u05F2\u0620-\u063F\u0641-\u064A\u066E-\u066F\u0671-\u06D3\u06D5\u06EE-\u06EF\u06FA-\u06FC\u06FF\u0710\u0712-\u072F\u074D-\u07A5\u07B1\u07CA-\u07EA\u0800-\u0815\u0840-\u0858\u08A0\u08A2-\u08AC\u0904-\u0939\u093D\u0950\u0958-\u0961\u0972-\u0977\u0979-\u097F\u0985-\u098C\u098F-\u0990\u0993-\u09A8\u09AA-\u09B0\u09B2\u09B6-\u09B9\u09BD\u09CE\u09DC-\u09DD\u09DF-\u09E1\u09F0-\u09F1\u0A05-\u0A0A\u0A0F-\u0A10\u0A13-\u0A28\u0A2A-\u0A30\u0A32-\u0A33\u0A35-\u0A36\u0A38-\u0A39\u0A59-\u0A5C\u0A5E\u0A72-\u0A74\u0A85-\u0A8D\u0A8F-\u0A91\u0A93-\u0AA8\u0AAA-\u0AB0\u0AB2-\u0AB3\u0AB5-\u0AB9\u0ABD\u0AD0\u0AE0-\u0AE1\u0B05-\u0B0C\u0B0F-\u0B10\u0B13-\u0B28\u0B2A-\u0B30\u0B32-\u0B33\u0B35-\u0B39\u0B3D\u0B5C-\u0B5D\u0B5F-\u0B61\u0B71\u0B83\u0B85-\u0B8A\u0B8E-\u0B90\u0B92-\u0B95\u0B99-\u0B9A\u0B9C\u0B9E-\u0B9F\u0BA3-\u0BA4\u0BA8-\u0BAA\u0BAE-\u0BB9\u0BD0\u0C05-\u0C0C\u0C0E-\u0C10\u0C12-\u0C28\u0C2A-\u0C33\u0C35-\u0C39\u0C3D\u0C58-\u0C59\u0C60-\u0C61\u0C85-\u0C8C\u0C8E-\u0C90\u0C92-\u0CA8\u0CAA-\u0CB3\u0CB5-\u0CB9\u0CBD\u0CDE\u0CE0-\u0CE1\u0CF1-\u0CF2\u0D05-\u0D0C\u0D0E-\u0D10\u0D12-\u0D3A\u0D3D\u0D4E\u0D60-\u0D61\u0D7A-\u0D7F\u0D85-\u0D96\u0D9A-\u0DB1\u0DB3-\u0DBB\u0DBD\u0DC0-\u0DC6\u0E01-\u0E30\u0E32-\u0E33\u0E40-\u0E45\u0E81-\u0E82\u0E84\u0E87-\u0E88\u0E8A\u0E8D\u0E94-\u0E97\u0E99-\u0E9F\u0EA1-\u0EA3\u0EA5\u0EA7\u0EAA-\u0EAB\u0EAD-\u0EB0\u0EB2-\u0EB3\u0EBD\u0EC0-\u0EC4\u0EDC-\u0EDF\u0F00\u0F40-\u0F47\u0F49-\u0F6C\u0F88-\u0F8C\u1000-\u102A\u103F\u1050-\u1055\u105A-\u105D\u1061\u1065-\u1066\u106E-\u1070\u1075-\u1081\u108E\u10D0-\u10FA\u10FD-\u1248\u124A-\u124D\u1250-\u1256\u1258\u125A-\u125D\u1260-\u1288\u128A-\u128D\u1290-\u12B0\u12B2-\u12B5\u12B8-\u12BE\u12C0\u12C2-\u12C5\u12C8-\u12D6\u12D8-\u1310\u1312-\u1315\u1318-\u135A\u1380-\u138F\u13A0-\u13F4\u1401-\u166C\u166F-\u167F\u1681-\u169A\u16A0-\u16EA\u1700-\u170C\u170E-\u1711\u1720-\u1731\u1740-\u1751\u1760-\u176C\u176E-\u1770\u1780-\u17B3\u17DC\u1820-\u1842\u1844-\u1877\u1880-\u18A8\u18AA\u18B0-\u18F5\u1900-\u191C\u1950-\u196D\u1970-\u1974\u1980-\u19AB\u19C1-\u19C7\u1A00-\u1A16\u1A20-\u1A54\u1B05-\u1B33\u1B45-\u1B4B\u1B83-\u1BA0\u1BAE-\u1BAF\u1BBA-\u1BE5\u1C00-\u1C23\u1C4D-\u1C4F\u1C5A-\u1C77\u1CE9-\u1CEC\u1CEE-\u1CF1\u1CF5-\u1CF6\u2135-\u2138\u2D30-\u2D67\u2D80-\u2D96\u2DA0-\u2DA6\u2DA8-\u2DAE\u2DB0-\u2DB6\u2DB8-\u2DBE\u2DC0-\u2DC6\u2DC8-\u2DCE\u2DD0-\u2DD6\u2DD8-\u2DDE\u3006\u303C\u3041-\u3096\u309F\u30A1-\u30FA\u30FF\u3105-\u312D\u3131-\u318E\u31A0-\u31BA\u31F0-\u31FF\u3400-\u4DB5\u4E00-\u9FCC\uA000-\uA014\uA016-\uA48C\uA4D0-\uA4F7\uA500-\uA60B\uA610-\uA61F\uA62A-\uA62B\uA66E\uA6A0-\uA6E5\uA7FB-\uA801\uA803-\uA805\uA807-\uA80A\uA80C-\uA822\uA840-\uA873\uA882-\uA8B3\uA8F2-\uA8F7\uA8FB\uA90A-\uA925\uA930-\uA946\uA960-\uA97C\uA984-\uA9B2\uAA00-\uAA28\uAA40-\uAA42\uAA44-\uAA4B\uAA60-\uAA6F\uAA71-\uAA76\uAA7A\uAA80-\uAAAF\uAAB1\uAAB5-\uAAB6\uAAB9-\uAABD\uAAC0\uAAC2\uAADB-\uAADC\uAAE0-\uAAEA\uAAF2\uAB01-\uAB06\uAB09-\uAB0E\uAB11-\uAB16\uAB20-\uAB26\uAB28-\uAB2E\uABC0-\uABE2\uAC00-\uD7A3\uD7B0-\uD7C6\uD7CB-\uD7FB\uF900-\uFA6D\uFA70-\uFAD9\uFB1D\uFB1F-\uFB28\uFB2A-\uFB36\uFB38-\uFB3C\uFB3E\uFB40-\uFB41\uFB43-\uFB44\uFB46-\uFBB1\uFBD3-\uFD3D\uFD50-\uFD8F\uFD92-\uFDC7\uFDF0-\uFDFB\uFE70-\uFE74\uFE76-\uFEFC\uFF66-\uFF6F\uFF71-\uFF9D\uFFA0-\uFFBE\uFFC2-\uFFC7\uFFCA-\uFFCF\uFFD2-\uFFD7\uFFDA-\uFFDC]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\xAA\\xBA\\u01BB\\u01C0-\\u01C3\\u0294\\u05D0-\\u05EA\\u05F0-\\u05F2\\u0620-\\u063F\\u0641-\\u064A\\u066E-\\u066F\\u0671-\\u06D3\\u06D5\\u06EE-\\u06EF\\u06FA-\\u06FC\\u06FF\\u0710\\u0712-\\u072F\\u074D-\\u07A5\\u07B1\\u07CA-\\u07EA\\u0800-\\u0815\\u0840-\\u0858\\u08A0\\u08A2-\\u08AC\\u0904-\\u0939\\u093D\\u0950\\u0958-\\u0961\\u0972-\\u0977\\u0979-\\u097F\\u0985-\\u098C\\u098F-\\u0990\\u0993-\\u09A8\\u09AA-\\u09B0\\u09B2\\u09B6-\\u09B9\\u09BD\\u09CE\\u09DC-\\u09DD\\u09DF-\\u09E1\\u09F0-\\u09F1\\u0A05-\\u0A0A\\u0A0F-\\u0A10\\u0A13-\\u0A28\\u0A2A-\\u0A30\\u0A32-\\u0A33\\u0A35-\\u0A36\\u0A38-\\u0A39\\u0A59-\\u0A5C\\u0A5E\\u0A72-\\u0A74\\u0A85-\\u0A8D\\u0A8F-\\u0A91\\u0A93-\\u0AA8\\u0AAA-\\u0AB0\\u0AB2-\\u0AB3\\u0AB5-\\u0AB9\\u0ABD\\u0AD0\\u0AE0-\\u0AE1\\u0B05-\\u0B0C\\u0B0F-\\u0B10\\u0B13-\\u0B28\\u0B2A-\\u0B30\\u0B32-\\u0B33\\u0B35-\\u0B39\\u0B3D\\u0B5C-\\u0B5D\\u0B5F-\\u0B61\\u0B71\\u0B83\\u0B85-\\u0B8A\\u0B8E-\\u0B90\\u0B92-\\u0B95\\u0B99-\\u0B9A\\u0B9C\\u0B9E-\\u0B9F\\u0BA3-\\u0BA4\\u0BA8-\\u0BAA\\u0BAE-\\u0BB9\\u0BD0\\u0C05-\\u0C0C\\u0C0E-\\u0C10\\u0C12-\\u0C28\\u0C2A-\\u0C33\\u0C35-\\u0C39\\u0C3D\\u0C58-\\u0C59\\u0C60-\\u0C61\\u0C85-\\u0C8C\\u0C8E-\\u0C90\\u0C92-\\u0CA8\\u0CAA-\\u0CB3\\u0CB5-\\u0CB9\\u0CBD\\u0CDE\\u0CE0-\\u0CE1\\u0CF1-\\u0CF2\\u0D05-\\u0D0C\\u0D0E-\\u0D10\\u0D12-\\u0D3A\\u0D3D\\u0D4E\\u0D60-\\u0D61\\u0D7A-\\u0D7F\\u0D85-\\u0D96\\u0D9A-\\u0DB1\\u0DB3-\\u0DBB\\u0DBD\\u0DC0-\\u0DC6\\u0E01-\\u0E30\\u0E32-\\u0E33\\u0E40-\\u0E45\\u0E81-\\u0E82\\u0E84\\u0E87-\\u0E88\\u0E8A\\u0E8D\\u0E94-\\u0E97\\u0E99-\\u0E9F\\u0EA1-\\u0EA3\\u0EA5\\u0EA7\\u0EAA-\\u0EAB\\u0EAD-\\u0EB0\\u0EB2-\\u0EB3\\u0EBD\\u0EC0-\\u0EC4\\u0EDC-\\u0EDF\\u0F00\\u0F40-\\u0F47\\u0F49-\\u0F6C\\u0F88-\\u0F8C\\u1000-\\u102A\\u103F\\u1050-\\u1055\\u105A-\\u105D\\u1061\\u1065-\\u1066\\u106E-\\u1070\\u1075-\\u1081\\u108E\\u10D0-\\u10FA\\u10FD-\\u1248\\u124A-\\u124D\\u1250-\\u1256\\u1258\\u125A-\\u125D\\u1260-\\u1288\\u128A-\\u128D\\u1290-\\u12B0\\u12B2-\\u12B5\\u12B8-\\u12BE\\u12C0\\u12C2-\\u12C5\\u12C8-\\u12D6\\u12D8-\\u1310\\u1312-\\u1315\\u1318-\\u135A\\u1380-\\u138F\\u13A0-\\u13F4\\u1401-\\u166C\\u166F-\\u167F\\u1681-\\u169A\\u16A0-\\u16EA\\u1700-\\u170C\\u170E-\\u1711\\u1720-\\u1731\\u1740-\\u1751\\u1760-\\u176C\\u176E-\\u1770\\u1780-\\u17B3\\u17DC\\u1820-\\u1842\\u1844-\\u1877\\u1880-\\u18A8\\u18AA\\u18B0-\\u18F5\\u1900-\\u191C\\u1950-\\u196D\\u1970-\\u1974\\u1980-\\u19AB\\u19C1-\\u19C7\\u1A00-\\u1A16\\u1A20-\\u1A54\\u1B05-\\u1B33\\u1B45-\\u1B4B\\u1B83-\\u1BA0\\u1BAE-\\u1BAF\\u1BBA-\\u1BE5\\u1C00-\\u1C23\\u1C4D-\\u1C4F\\u1C5A-\\u1C77\\u1CE9-\\u1CEC\\u1CEE-\\u1CF1\\u1CF5-\\u1CF6\\u2135-\\u2138\\u2D30-\\u2D67\\u2D80-\\u2D96\\u2DA0-\\u2DA6\\u2DA8-\\u2DAE\\u2DB0-\\u2DB6\\u2DB8-\\u2DBE\\u2DC0-\\u2DC6\\u2DC8-\\u2DCE\\u2DD0-\\u2DD6\\u2DD8-\\u2DDE\\u3006\\u303C\\u3041-\\u3096\\u309F\\u30A1-\\u30FA\\u30FF\\u3105-\\u312D\\u3131-\\u318E\\u31A0-\\u31BA\\u31F0-\\u31FF\\u3400-\\u4DB5\\u4E00-\\u9FCC\\uA000-\\uA014\\uA016-\\uA48C\\uA4D0-\\uA4F7\\uA500-\\uA60B\\uA610-\\uA61F\\uA62A-\\uA62B\\uA66E\\uA6A0-\\uA6E5\\uA7FB-\\uA801\\uA803-\\uA805\\uA807-\\uA80A\\uA80C-\\uA822\\uA840-\\uA873\\uA882-\\uA8B3\\uA8F2-\\uA8F7\\uA8FB\\uA90A-\\uA925\\uA930-\\uA946\\uA960-\\uA97C\\uA984-\\uA9B2\\uAA00-\\uAA28\\uAA40-\\uAA42\\uAA44-\\uAA4B\\uAA60-\\uAA6F\\uAA71-\\uAA76\\uAA7A\\uAA80-\\uAAAF\\uAAB1\\uAAB5-\\uAAB6\\uAAB9-\\uAABD\\uAAC0\\uAAC2\\uAADB-\\uAADC\\uAAE0-\\uAAEA\\uAAF2\\uAB01-\\uAB06\\uAB09-\\uAB0E\\uAB11-\\uAB16\\uAB20-\\uAB26\\uAB28-\\uAB2E\\uABC0-\\uABE2\\uAC00-\\uD7A3\\uD7B0-\\uD7C6\\uD7CB-\\uD7FB\\uF900-\\uFA6D\\uFA70-\\uFAD9\\uFB1D\\uFB1F-\\uFB28\\uFB2A-\\uFB36\\uFB38-\\uFB3C\\uFB3E\\uFB40-\\uFB41\\uFB43-\\uFB44\\uFB46-\\uFBB1\\uFBD3-\\uFD3D\\uFD50-\\uFD8F\\uFD92-\\uFDC7\\uFDF0-\\uFDFB\\uFE70-\\uFE74\\uFE76-\\uFEFC\\uFF66-\\uFF6F\\uFF71-\\uFF9D\\uFFA0-\\uFFBE\\uFFC2-\\uFFC7\\uFFCA-\\uFFCF\\uFFD2-\\uFFD7\\uFFDA-\\uFFDC]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Lt() {
-        var result0;
-        
-        if (/^[\u01C5\u01C8\u01CB\u01F2\u1F88-\u1F8F\u1F98-\u1F9F\u1FA8-\u1FAF\u1FBC\u1FCC\u1FFC]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\u01C5\\u01C8\\u01CB\\u01F2\\u1F88-\\u1F8F\\u1F98-\\u1F9F\\u1FA8-\\u1FAF\\u1FBC\\u1FCC\\u1FFC]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Lu() {
-        var result0;
-        
-        if (/^[A-Z\xC0-\xD6\xD8-\xDE\u0100\u0102\u0104\u0106\u0108\u010A\u010C\u010E\u0110\u0112\u0114\u0116\u0118\u011A\u011C\u011E\u0120\u0122\u0124\u0126\u0128\u012A\u012C\u012E\u0130\u0132\u0134\u0136\u0139\u013B\u013D\u013F\u0141\u0143\u0145\u0147\u014A\u014C\u014E\u0150\u0152\u0154\u0156\u0158\u015A\u015C\u015E\u0160\u0162\u0164\u0166\u0168\u016A\u016C\u016E\u0170\u0172\u0174\u0176\u0178-\u0179\u017B\u017D\u0181-\u0182\u0184\u0186-\u0187\u0189-\u018B\u018E-\u0191\u0193-\u0194\u0196-\u0198\u019C-\u019D\u019F-\u01A0\u01A2\u01A4\u01A6-\u01A7\u01A9\u01AC\u01AE-\u01AF\u01B1-\u01B3\u01B5\u01B7-\u01B8\u01BC\u01C4\u01C7\u01CA\u01CD\u01CF\u01D1\u01D3\u01D5\u01D7\u01D9\u01DB\u01DE\u01E0\u01E2\u01E4\u01E6\u01E8\u01EA\u01EC\u01EE\u01F1\u01F4\u01F6-\u01F8\u01FA\u01FC\u01FE\u0200\u0202\u0204\u0206\u0208\u020A\u020C\u020E\u0210\u0212\u0214\u0216\u0218\u021A\u021C\u021E\u0220\u0222\u0224\u0226\u0228\u022A\u022C\u022E\u0230\u0232\u023A-\u023B\u023D-\u023E\u0241\u0243-\u0246\u0248\u024A\u024C\u024E\u0370\u0372\u0376\u0386\u0388-\u038A\u038C\u038E-\u038F\u0391-\u03A1\u03A3-\u03AB\u03CF\u03D2-\u03D4\u03D8\u03DA\u03DC\u03DE\u03E0\u03E2\u03E4\u03E6\u03E8\u03EA\u03EC\u03EE\u03F4\u03F7\u03F9-\u03FA\u03FD-\u042F\u0460\u0462\u0464\u0466\u0468\u046A\u046C\u046E\u0470\u0472\u0474\u0476\u0478\u047A\u047C\u047E\u0480\u048A\u048C\u048E\u0490\u0492\u0494\u0496\u0498\u049A\u049C\u049E\u04A0\u04A2\u04A4\u04A6\u04A8\u04AA\u04AC\u04AE\u04B0\u04B2\u04B4\u04B6\u04B8\u04BA\u04BC\u04BE\u04C0-\u04C1\u04C3\u04C5\u04C7\u04C9\u04CB\u04CD\u04D0\u04D2\u04D4\u04D6\u04D8\u04DA\u04DC\u04DE\u04E0\u04E2\u04E4\u04E6\u04E8\u04EA\u04EC\u04EE\u04F0\u04F2\u04F4\u04F6\u04F8\u04FA\u04FC\u04FE\u0500\u0502\u0504\u0506\u0508\u050A\u050C\u050E\u0510\u0512\u0514\u0516\u0518\u051A\u051C\u051E\u0520\u0522\u0524\u0526\u0531-\u0556\u10A0-\u10C5\u10C7\u10CD\u1E00\u1E02\u1E04\u1E06\u1E08\u1E0A\u1E0C\u1E0E\u1E10\u1E12\u1E14\u1E16\u1E18\u1E1A\u1E1C\u1E1E\u1E20\u1E22\u1E24\u1E26\u1E28\u1E2A\u1E2C\u1E2E\u1E30\u1E32\u1E34\u1E36\u1E38\u1E3A\u1E3C\u1E3E\u1E40\u1E42\u1E44\u1E46\u1E48\u1E4A\u1E4C\u1E4E\u1E50\u1E52\u1E54\u1E56\u1E58\u1E5A\u1E5C\u1E5E\u1E60\u1E62\u1E64\u1E66\u1E68\u1E6A\u1E6C\u1E6E\u1E70\u1E72\u1E74\u1E76\u1E78\u1E7A\u1E7C\u1E7E\u1E80\u1E82\u1E84\u1E86\u1E88\u1E8A\u1E8C\u1E8E\u1E90\u1E92\u1E94\u1E9E\u1EA0\u1EA2\u1EA4\u1EA6\u1EA8\u1EAA\u1EAC\u1EAE\u1EB0\u1EB2\u1EB4\u1EB6\u1EB8\u1EBA\u1EBC\u1EBE\u1EC0\u1EC2\u1EC4\u1EC6\u1EC8\u1ECA\u1ECC\u1ECE\u1ED0\u1ED2\u1ED4\u1ED6\u1ED8\u1EDA\u1EDC\u1EDE\u1EE0\u1EE2\u1EE4\u1EE6\u1EE8\u1EEA\u1EEC\u1EEE\u1EF0\u1EF2\u1EF4\u1EF6\u1EF8\u1EFA\u1EFC\u1EFE\u1F08-\u1F0F\u1F18-\u1F1D\u1F28-\u1F2F\u1F38-\u1F3F\u1F48-\u1F4D\u1F59\u1F5B\u1F5D\u1F5F\u1F68-\u1F6F\u1FB8-\u1FBB\u1FC8-\u1FCB\u1FD8-\u1FDB\u1FE8-\u1FEC\u1FF8-\u1FFB\u2102\u2107\u210B-\u210D\u2110-\u2112\u2115\u2119-\u211D\u2124\u2126\u2128\u212A-\u212D\u2130-\u2133\u213E-\u213F\u2145\u2183\u2C00-\u2C2E\u2C60\u2C62-\u2C64\u2C67\u2C69\u2C6B\u2C6D-\u2C70\u2C72\u2C75\u2C7E-\u2C80\u2C82\u2C84\u2C86\u2C88\u2C8A\u2C8C\u2C8E\u2C90\u2C92\u2C94\u2C96\u2C98\u2C9A\u2C9C\u2C9E\u2CA0\u2CA2\u2CA4\u2CA6\u2CA8\u2CAA\u2CAC\u2CAE\u2CB0\u2CB2\u2CB4\u2CB6\u2CB8\u2CBA\u2CBC\u2CBE\u2CC0\u2CC2\u2CC4\u2CC6\u2CC8\u2CCA\u2CCC\u2CCE\u2CD0\u2CD2\u2CD4\u2CD6\u2CD8\u2CDA\u2CDC\u2CDE\u2CE0\u2CE2\u2CEB\u2CED\u2CF2\uA640\uA642\uA644\uA646\uA648\uA64A\uA64C\uA64E\uA650\uA652\uA654\uA656\uA658\uA65A\uA65C\uA65E\uA660\uA662\uA664\uA666\uA668\uA66A\uA66C\uA680\uA682\uA684\uA686\uA688\uA68A\uA68C\uA68E\uA690\uA692\uA694\uA696\uA722\uA724\uA726\uA728\uA72A\uA72C\uA72E\uA732\uA734\uA736\uA738\uA73A\uA73C\uA73E\uA740\uA742\uA744\uA746\uA748\uA74A\uA74C\uA74E\uA750\uA752\uA754\uA756\uA758\uA75A\uA75C\uA75E\uA760\uA762\uA764\uA766\uA768\uA76A\uA76C\uA76E\uA779\uA77B\uA77D-\uA77E\uA780\uA782\uA784\uA786\uA78B\uA78D\uA790\uA792\uA7A0\uA7A2\uA7A4\uA7A6\uA7A8\uA7AA\uFF21-\uFF3A]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[A-Z\\xC0-\\xD6\\xD8-\\xDE\\u0100\\u0102\\u0104\\u0106\\u0108\\u010A\\u010C\\u010E\\u0110\\u0112\\u0114\\u0116\\u0118\\u011A\\u011C\\u011E\\u0120\\u0122\\u0124\\u0126\\u0128\\u012A\\u012C\\u012E\\u0130\\u0132\\u0134\\u0136\\u0139\\u013B\\u013D\\u013F\\u0141\\u0143\\u0145\\u0147\\u014A\\u014C\\u014E\\u0150\\u0152\\u0154\\u0156\\u0158\\u015A\\u015C\\u015E\\u0160\\u0162\\u0164\\u0166\\u0168\\u016A\\u016C\\u016E\\u0170\\u0172\\u0174\\u0176\\u0178-\\u0179\\u017B\\u017D\\u0181-\\u0182\\u0184\\u0186-\\u0187\\u0189-\\u018B\\u018E-\\u0191\\u0193-\\u0194\\u0196-\\u0198\\u019C-\\u019D\\u019F-\\u01A0\\u01A2\\u01A4\\u01A6-\\u01A7\\u01A9\\u01AC\\u01AE-\\u01AF\\u01B1-\\u01B3\\u01B5\\u01B7-\\u01B8\\u01BC\\u01C4\\u01C7\\u01CA\\u01CD\\u01CF\\u01D1\\u01D3\\u01D5\\u01D7\\u01D9\\u01DB\\u01DE\\u01E0\\u01E2\\u01E4\\u01E6\\u01E8\\u01EA\\u01EC\\u01EE\\u01F1\\u01F4\\u01F6-\\u01F8\\u01FA\\u01FC\\u01FE\\u0200\\u0202\\u0204\\u0206\\u0208\\u020A\\u020C\\u020E\\u0210\\u0212\\u0214\\u0216\\u0218\\u021A\\u021C\\u021E\\u0220\\u0222\\u0224\\u0226\\u0228\\u022A\\u022C\\u022E\\u0230\\u0232\\u023A-\\u023B\\u023D-\\u023E\\u0241\\u0243-\\u0246\\u0248\\u024A\\u024C\\u024E\\u0370\\u0372\\u0376\\u0386\\u0388-\\u038A\\u038C\\u038E-\\u038F\\u0391-\\u03A1\\u03A3-\\u03AB\\u03CF\\u03D2-\\u03D4\\u03D8\\u03DA\\u03DC\\u03DE\\u03E0\\u03E2\\u03E4\\u03E6\\u03E8\\u03EA\\u03EC\\u03EE\\u03F4\\u03F7\\u03F9-\\u03FA\\u03FD-\\u042F\\u0460\\u0462\\u0464\\u0466\\u0468\\u046A\\u046C\\u046E\\u0470\\u0472\\u0474\\u0476\\u0478\\u047A\\u047C\\u047E\\u0480\\u048A\\u048C\\u048E\\u0490\\u0492\\u0494\\u0496\\u0498\\u049A\\u049C\\u049E\\u04A0\\u04A2\\u04A4\\u04A6\\u04A8\\u04AA\\u04AC\\u04AE\\u04B0\\u04B2\\u04B4\\u04B6\\u04B8\\u04BA\\u04BC\\u04BE\\u04C0-\\u04C1\\u04C3\\u04C5\\u04C7\\u04C9\\u04CB\\u04CD\\u04D0\\u04D2\\u04D4\\u04D6\\u04D8\\u04DA\\u04DC\\u04DE\\u04E0\\u04E2\\u04E4\\u04E6\\u04E8\\u04EA\\u04EC\\u04EE\\u04F0\\u04F2\\u04F4\\u04F6\\u04F8\\u04FA\\u04FC\\u04FE\\u0500\\u0502\\u0504\\u0506\\u0508\\u050A\\u050C\\u050E\\u0510\\u0512\\u0514\\u0516\\u0518\\u051A\\u051C\\u051E\\u0520\\u0522\\u0524\\u0526\\u0531-\\u0556\\u10A0-\\u10C5\\u10C7\\u10CD\\u1E00\\u1E02\\u1E04\\u1E06\\u1E08\\u1E0A\\u1E0C\\u1E0E\\u1E10\\u1E12\\u1E14\\u1E16\\u1E18\\u1E1A\\u1E1C\\u1E1E\\u1E20\\u1E22\\u1E24\\u1E26\\u1E28\\u1E2A\\u1E2C\\u1E2E\\u1E30\\u1E32\\u1E34\\u1E36\\u1E38\\u1E3A\\u1E3C\\u1E3E\\u1E40\\u1E42\\u1E44\\u1E46\\u1E48\\u1E4A\\u1E4C\\u1E4E\\u1E50\\u1E52\\u1E54\\u1E56\\u1E58\\u1E5A\\u1E5C\\u1E5E\\u1E60\\u1E62\\u1E64\\u1E66\\u1E68\\u1E6A\\u1E6C\\u1E6E\\u1E70\\u1E72\\u1E74\\u1E76\\u1E78\\u1E7A\\u1E7C\\u1E7E\\u1E80\\u1E82\\u1E84\\u1E86\\u1E88\\u1E8A\\u1E8C\\u1E8E\\u1E90\\u1E92\\u1E94\\u1E9E\\u1EA0\\u1EA2\\u1EA4\\u1EA6\\u1EA8\\u1EAA\\u1EAC\\u1EAE\\u1EB0\\u1EB2\\u1EB4\\u1EB6\\u1EB8\\u1EBA\\u1EBC\\u1EBE\\u1EC0\\u1EC2\\u1EC4\\u1EC6\\u1EC8\\u1ECA\\u1ECC\\u1ECE\\u1ED0\\u1ED2\\u1ED4\\u1ED6\\u1ED8\\u1EDA\\u1EDC\\u1EDE\\u1EE0\\u1EE2\\u1EE4\\u1EE6\\u1EE8\\u1EEA\\u1EEC\\u1EEE\\u1EF0\\u1EF2\\u1EF4\\u1EF6\\u1EF8\\u1EFA\\u1EFC\\u1EFE\\u1F08-\\u1F0F\\u1F18-\\u1F1D\\u1F28-\\u1F2F\\u1F38-\\u1F3F\\u1F48-\\u1F4D\\u1F59\\u1F5B\\u1F5D\\u1F5F\\u1F68-\\u1F6F\\u1FB8-\\u1FBB\\u1FC8-\\u1FCB\\u1FD8-\\u1FDB\\u1FE8-\\u1FEC\\u1FF8-\\u1FFB\\u2102\\u2107\\u210B-\\u210D\\u2110-\\u2112\\u2115\\u2119-\\u211D\\u2124\\u2126\\u2128\\u212A-\\u212D\\u2130-\\u2133\\u213E-\\u213F\\u2145\\u2183\\u2C00-\\u2C2E\\u2C60\\u2C62-\\u2C64\\u2C67\\u2C69\\u2C6B\\u2C6D-\\u2C70\\u2C72\\u2C75\\u2C7E-\\u2C80\\u2C82\\u2C84\\u2C86\\u2C88\\u2C8A\\u2C8C\\u2C8E\\u2C90\\u2C92\\u2C94\\u2C96\\u2C98\\u2C9A\\u2C9C\\u2C9E\\u2CA0\\u2CA2\\u2CA4\\u2CA6\\u2CA8\\u2CAA\\u2CAC\\u2CAE\\u2CB0\\u2CB2\\u2CB4\\u2CB6\\u2CB8\\u2CBA\\u2CBC\\u2CBE\\u2CC0\\u2CC2\\u2CC4\\u2CC6\\u2CC8\\u2CCA\\u2CCC\\u2CCE\\u2CD0\\u2CD2\\u2CD4\\u2CD6\\u2CD8\\u2CDA\\u2CDC\\u2CDE\\u2CE0\\u2CE2\\u2CEB\\u2CED\\u2CF2\\uA640\\uA642\\uA644\\uA646\\uA648\\uA64A\\uA64C\\uA64E\\uA650\\uA652\\uA654\\uA656\\uA658\\uA65A\\uA65C\\uA65E\\uA660\\uA662\\uA664\\uA666\\uA668\\uA66A\\uA66C\\uA680\\uA682\\uA684\\uA686\\uA688\\uA68A\\uA68C\\uA68E\\uA690\\uA692\\uA694\\uA696\\uA722\\uA724\\uA726\\uA728\\uA72A\\uA72C\\uA72E\\uA732\\uA734\\uA736\\uA738\\uA73A\\uA73C\\uA73E\\uA740\\uA742\\uA744\\uA746\\uA748\\uA74A\\uA74C\\uA74E\\uA750\\uA752\\uA754\\uA756\\uA758\\uA75A\\uA75C\\uA75E\\uA760\\uA762\\uA764\\uA766\\uA768\\uA76A\\uA76C\\uA76E\\uA779\\uA77B\\uA77D-\\uA77E\\uA780\\uA782\\uA784\\uA786\\uA78B\\uA78D\\uA790\\uA792\\uA7A0\\uA7A2\\uA7A4\\uA7A6\\uA7A8\\uA7AA\\uFF21-\\uFF3A]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Mc() {
-        var result0;
-        
-        if (/^[\u0903\u093B\u093E-\u0940\u0949-\u094C\u094E-\u094F\u0982-\u0983\u09BE-\u09C0\u09C7-\u09C8\u09CB-\u09CC\u09D7\u0A03\u0A3E-\u0A40\u0A83\u0ABE-\u0AC0\u0AC9\u0ACB-\u0ACC\u0B02-\u0B03\u0B3E\u0B40\u0B47-\u0B48\u0B4B-\u0B4C\u0B57\u0BBE-\u0BBF\u0BC1-\u0BC2\u0BC6-\u0BC8\u0BCA-\u0BCC\u0BD7\u0C01-\u0C03\u0C41-\u0C44\u0C82-\u0C83\u0CBE\u0CC0-\u0CC4\u0CC7-\u0CC8\u0CCA-\u0CCB\u0CD5-\u0CD6\u0D02-\u0D03\u0D3E-\u0D40\u0D46-\u0D48\u0D4A-\u0D4C\u0D57\u0D82-\u0D83\u0DCF-\u0DD1\u0DD8-\u0DDF\u0DF2-\u0DF3\u0F3E-\u0F3F\u0F7F\u102B-\u102C\u1031\u1038\u103B-\u103C\u1056-\u1057\u1062-\u1064\u1067-\u106D\u1083-\u1084\u1087-\u108C\u108F\u109A-\u109C\u17B6\u17BE-\u17C5\u17C7-\u17C8\u1923-\u1926\u1929-\u192B\u1930-\u1931\u1933-\u1938\u19B0-\u19C0\u19C8-\u19C9\u1A19-\u1A1A\u1A55\u1A57\u1A61\u1A63-\u1A64\u1A6D-\u1A72\u1B04\u1B35\u1B3B\u1B3D-\u1B41\u1B43-\u1B44\u1B82\u1BA1\u1BA6-\u1BA7\u1BAA\u1BAC-\u1BAD\u1BE7\u1BEA-\u1BEC\u1BEE\u1BF2-\u1BF3\u1C24-\u1C2B\u1C34-\u1C35\u1CE1\u1CF2-\u1CF3\u302E-\u302F\uA823-\uA824\uA827\uA880-\uA881\uA8B4-\uA8C3\uA952-\uA953\uA983\uA9B4-\uA9B5\uA9BA-\uA9BB\uA9BD-\uA9C0\uAA2F-\uAA30\uAA33-\uAA34\uAA4D\uAA7B\uAAEB\uAAEE-\uAAEF\uAAF5\uABE3-\uABE4\uABE6-\uABE7\uABE9-\uABEA\uABEC]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\u0903\\u093B\\u093E-\\u0940\\u0949-\\u094C\\u094E-\\u094F\\u0982-\\u0983\\u09BE-\\u09C0\\u09C7-\\u09C8\\u09CB-\\u09CC\\u09D7\\u0A03\\u0A3E-\\u0A40\\u0A83\\u0ABE-\\u0AC0\\u0AC9\\u0ACB-\\u0ACC\\u0B02-\\u0B03\\u0B3E\\u0B40\\u0B47-\\u0B48\\u0B4B-\\u0B4C\\u0B57\\u0BBE-\\u0BBF\\u0BC1-\\u0BC2\\u0BC6-\\u0BC8\\u0BCA-\\u0BCC\\u0BD7\\u0C01-\\u0C03\\u0C41-\\u0C44\\u0C82-\\u0C83\\u0CBE\\u0CC0-\\u0CC4\\u0CC7-\\u0CC8\\u0CCA-\\u0CCB\\u0CD5-\\u0CD6\\u0D02-\\u0D03\\u0D3E-\\u0D40\\u0D46-\\u0D48\\u0D4A-\\u0D4C\\u0D57\\u0D82-\\u0D83\\u0DCF-\\u0DD1\\u0DD8-\\u0DDF\\u0DF2-\\u0DF3\\u0F3E-\\u0F3F\\u0F7F\\u102B-\\u102C\\u1031\\u1038\\u103B-\\u103C\\u1056-\\u1057\\u1062-\\u1064\\u1067-\\u106D\\u1083-\\u1084\\u1087-\\u108C\\u108F\\u109A-\\u109C\\u17B6\\u17BE-\\u17C5\\u17C7-\\u17C8\\u1923-\\u1926\\u1929-\\u192B\\u1930-\\u1931\\u1933-\\u1938\\u19B0-\\u19C0\\u19C8-\\u19C9\\u1A19-\\u1A1A\\u1A55\\u1A57\\u1A61\\u1A63-\\u1A64\\u1A6D-\\u1A72\\u1B04\\u1B35\\u1B3B\\u1B3D-\\u1B41\\u1B43-\\u1B44\\u1B82\\u1BA1\\u1BA6-\\u1BA7\\u1BAA\\u1BAC-\\u1BAD\\u1BE7\\u1BEA-\\u1BEC\\u1BEE\\u1BF2-\\u1BF3\\u1C24-\\u1C2B\\u1C34-\\u1C35\\u1CE1\\u1CF2-\\u1CF3\\u302E-\\u302F\\uA823-\\uA824\\uA827\\uA880-\\uA881\\uA8B4-\\uA8C3\\uA952-\\uA953\\uA983\\uA9B4-\\uA9B5\\uA9BA-\\uA9BB\\uA9BD-\\uA9C0\\uAA2F-\\uAA30\\uAA33-\\uAA34\\uAA4D\\uAA7B\\uAAEB\\uAAEE-\\uAAEF\\uAAF5\\uABE3-\\uABE4\\uABE6-\\uABE7\\uABE9-\\uABEA\\uABEC]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Mn() {
-        var result0;
-        
-        if (/^[\u0300-\u036F\u0483-\u0487\u0591-\u05BD\u05BF\u05C1-\u05C2\u05C4-\u05C5\u05C7\u0610-\u061A\u064B-\u065F\u0670\u06D6-\u06DC\u06DF-\u06E4\u06E7-\u06E8\u06EA-\u06ED\u0711\u0730-\u074A\u07A6-\u07B0\u07EB-\u07F3\u0816-\u0819\u081B-\u0823\u0825-\u0827\u0829-\u082D\u0859-\u085B\u08E4-\u08FE\u0900-\u0902\u093A\u093C\u0941-\u0948\u094D\u0951-\u0957\u0962-\u0963\u0981\u09BC\u09C1-\u09C4\u09CD\u09E2-\u09E3\u0A01-\u0A02\u0A3C\u0A41-\u0A42\u0A47-\u0A48\u0A4B-\u0A4D\u0A51\u0A70-\u0A71\u0A75\u0A81-\u0A82\u0ABC\u0AC1-\u0AC5\u0AC7-\u0AC8\u0ACD\u0AE2-\u0AE3\u0B01\u0B3C\u0B3F\u0B41-\u0B44\u0B4D\u0B56\u0B62-\u0B63\u0B82\u0BC0\u0BCD\u0C3E-\u0C40\u0C46-\u0C48\u0C4A-\u0C4D\u0C55-\u0C56\u0C62-\u0C63\u0CBC\u0CBF\u0CC6\u0CCC-\u0CCD\u0CE2-\u0CE3\u0D41-\u0D44\u0D4D\u0D62-\u0D63\u0DCA\u0DD2-\u0DD4\u0DD6\u0E31\u0E34-\u0E3A\u0E47-\u0E4E\u0EB1\u0EB4-\u0EB9\u0EBB-\u0EBC\u0EC8-\u0ECD\u0F18-\u0F19\u0F35\u0F37\u0F39\u0F71-\u0F7E\u0F80-\u0F84\u0F86-\u0F87\u0F8D-\u0F97\u0F99-\u0FBC\u0FC6\u102D-\u1030\u1032-\u1037\u1039-\u103A\u103D-\u103E\u1058-\u1059\u105E-\u1060\u1071-\u1074\u1082\u1085-\u1086\u108D\u109D\u135D-\u135F\u1712-\u1714\u1732-\u1734\u1752-\u1753\u1772-\u1773\u17B4-\u17B5\u17B7-\u17BD\u17C6\u17C9-\u17D3\u17DD\u180B-\u180D\u18A9\u1920-\u1922\u1927-\u1928\u1932\u1939-\u193B\u1A17-\u1A18\u1A1B\u1A56\u1A58-\u1A5E\u1A60\u1A62\u1A65-\u1A6C\u1A73-\u1A7C\u1A7F\u1B00-\u1B03\u1B34\u1B36-\u1B3A\u1B3C\u1B42\u1B6B-\u1B73\u1B80-\u1B81\u1BA2-\u1BA5\u1BA8-\u1BA9\u1BAB\u1BE6\u1BE8-\u1BE9\u1BED\u1BEF-\u1BF1\u1C2C-\u1C33\u1C36-\u1C37\u1CD0-\u1CD2\u1CD4-\u1CE0\u1CE2-\u1CE8\u1CED\u1CF4\u1DC0-\u1DE6\u1DFC-\u1DFF\u20D0-\u20DC\u20E1\u20E5-\u20F0\u2CEF-\u2CF1\u2D7F\u2DE0-\u2DFF\u302A-\u302D\u3099-\u309A\uA66F\uA674-\uA67D\uA69F\uA6F0-\uA6F1\uA802\uA806\uA80B\uA825-\uA826\uA8C4\uA8E0-\uA8F1\uA926-\uA92D\uA947-\uA951\uA980-\uA982\uA9B3\uA9B6-\uA9B9\uA9BC\uAA29-\uAA2E\uAA31-\uAA32\uAA35-\uAA36\uAA43\uAA4C\uAAB0\uAAB2-\uAAB4\uAAB7-\uAAB8\uAABE-\uAABF\uAAC1\uAAEC-\uAAED\uAAF6\uABE5\uABE8\uABED\uFB1E\uFE00-\uFE0F\uFE20-\uFE26]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\u0300-\\u036F\\u0483-\\u0487\\u0591-\\u05BD\\u05BF\\u05C1-\\u05C2\\u05C4-\\u05C5\\u05C7\\u0610-\\u061A\\u064B-\\u065F\\u0670\\u06D6-\\u06DC\\u06DF-\\u06E4\\u06E7-\\u06E8\\u06EA-\\u06ED\\u0711\\u0730-\\u074A\\u07A6-\\u07B0\\u07EB-\\u07F3\\u0816-\\u0819\\u081B-\\u0823\\u0825-\\u0827\\u0829-\\u082D\\u0859-\\u085B\\u08E4-\\u08FE\\u0900-\\u0902\\u093A\\u093C\\u0941-\\u0948\\u094D\\u0951-\\u0957\\u0962-\\u0963\\u0981\\u09BC\\u09C1-\\u09C4\\u09CD\\u09E2-\\u09E3\\u0A01-\\u0A02\\u0A3C\\u0A41-\\u0A42\\u0A47-\\u0A48\\u0A4B-\\u0A4D\\u0A51\\u0A70-\\u0A71\\u0A75\\u0A81-\\u0A82\\u0ABC\\u0AC1-\\u0AC5\\u0AC7-\\u0AC8\\u0ACD\\u0AE2-\\u0AE3\\u0B01\\u0B3C\\u0B3F\\u0B41-\\u0B44\\u0B4D\\u0B56\\u0B62-\\u0B63\\u0B82\\u0BC0\\u0BCD\\u0C3E-\\u0C40\\u0C46-\\u0C48\\u0C4A-\\u0C4D\\u0C55-\\u0C56\\u0C62-\\u0C63\\u0CBC\\u0CBF\\u0CC6\\u0CCC-\\u0CCD\\u0CE2-\\u0CE3\\u0D41-\\u0D44\\u0D4D\\u0D62-\\u0D63\\u0DCA\\u0DD2-\\u0DD4\\u0DD6\\u0E31\\u0E34-\\u0E3A\\u0E47-\\u0E4E\\u0EB1\\u0EB4-\\u0EB9\\u0EBB-\\u0EBC\\u0EC8-\\u0ECD\\u0F18-\\u0F19\\u0F35\\u0F37\\u0F39\\u0F71-\\u0F7E\\u0F80-\\u0F84\\u0F86-\\u0F87\\u0F8D-\\u0F97\\u0F99-\\u0FBC\\u0FC6\\u102D-\\u1030\\u1032-\\u1037\\u1039-\\u103A\\u103D-\\u103E\\u1058-\\u1059\\u105E-\\u1060\\u1071-\\u1074\\u1082\\u1085-\\u1086\\u108D\\u109D\\u135D-\\u135F\\u1712-\\u1714\\u1732-\\u1734\\u1752-\\u1753\\u1772-\\u1773\\u17B4-\\u17B5\\u17B7-\\u17BD\\u17C6\\u17C9-\\u17D3\\u17DD\\u180B-\\u180D\\u18A9\\u1920-\\u1922\\u1927-\\u1928\\u1932\\u1939-\\u193B\\u1A17-\\u1A18\\u1A1B\\u1A56\\u1A58-\\u1A5E\\u1A60\\u1A62\\u1A65-\\u1A6C\\u1A73-\\u1A7C\\u1A7F\\u1B00-\\u1B03\\u1B34\\u1B36-\\u1B3A\\u1B3C\\u1B42\\u1B6B-\\u1B73\\u1B80-\\u1B81\\u1BA2-\\u1BA5\\u1BA8-\\u1BA9\\u1BAB\\u1BE6\\u1BE8-\\u1BE9\\u1BED\\u1BEF-\\u1BF1\\u1C2C-\\u1C33\\u1C36-\\u1C37\\u1CD0-\\u1CD2\\u1CD4-\\u1CE0\\u1CE2-\\u1CE8\\u1CED\\u1CF4\\u1DC0-\\u1DE6\\u1DFC-\\u1DFF\\u20D0-\\u20DC\\u20E1\\u20E5-\\u20F0\\u2CEF-\\u2CF1\\u2D7F\\u2DE0-\\u2DFF\\u302A-\\u302D\\u3099-\\u309A\\uA66F\\uA674-\\uA67D\\uA69F\\uA6F0-\\uA6F1\\uA802\\uA806\\uA80B\\uA825-\\uA826\\uA8C4\\uA8E0-\\uA8F1\\uA926-\\uA92D\\uA947-\\uA951\\uA980-\\uA982\\uA9B3\\uA9B6-\\uA9B9\\uA9BC\\uAA29-\\uAA2E\\uAA31-\\uAA32\\uAA35-\\uAA36\\uAA43\\uAA4C\\uAAB0\\uAAB2-\\uAAB4\\uAAB7-\\uAAB8\\uAABE-\\uAABF\\uAAC1\\uAAEC-\\uAAED\\uAAF6\\uABE5\\uABE8\\uABED\\uFB1E\\uFE00-\\uFE0F\\uFE20-\\uFE26]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Nd() {
-        var result0;
-        
-        if (/^[0-9\u0660-\u0669\u06F0-\u06F9\u07C0-\u07C9\u0966-\u096F\u09E6-\u09EF\u0A66-\u0A6F\u0AE6-\u0AEF\u0B66-\u0B6F\u0BE6-\u0BEF\u0C66-\u0C6F\u0CE6-\u0CEF\u0D66-\u0D6F\u0E50-\u0E59\u0ED0-\u0ED9\u0F20-\u0F29\u1040-\u1049\u1090-\u1099\u17E0-\u17E9\u1810-\u1819\u1946-\u194F\u19D0-\u19D9\u1A80-\u1A89\u1A90-\u1A99\u1B50-\u1B59\u1BB0-\u1BB9\u1C40-\u1C49\u1C50-\u1C59\uA620-\uA629\uA8D0-\uA8D9\uA900-\uA909\uA9D0-\uA9D9\uAA50-\uAA59\uABF0-\uABF9\uFF10-\uFF19]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[0-9\\u0660-\\u0669\\u06F0-\\u06F9\\u07C0-\\u07C9\\u0966-\\u096F\\u09E6-\\u09EF\\u0A66-\\u0A6F\\u0AE6-\\u0AEF\\u0B66-\\u0B6F\\u0BE6-\\u0BEF\\u0C66-\\u0C6F\\u0CE6-\\u0CEF\\u0D66-\\u0D6F\\u0E50-\\u0E59\\u0ED0-\\u0ED9\\u0F20-\\u0F29\\u1040-\\u1049\\u1090-\\u1099\\u17E0-\\u17E9\\u1810-\\u1819\\u1946-\\u194F\\u19D0-\\u19D9\\u1A80-\\u1A89\\u1A90-\\u1A99\\u1B50-\\u1B59\\u1BB0-\\u1BB9\\u1C40-\\u1C49\\u1C50-\\u1C59\\uA620-\\uA629\\uA8D0-\\uA8D9\\uA900-\\uA909\\uA9D0-\\uA9D9\\uAA50-\\uAA59\\uABF0-\\uABF9\\uFF10-\\uFF19]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Nl() {
-        var result0;
-        
-        if (/^[\u16EE-\u16F0\u2160-\u2182\u2185-\u2188\u3007\u3021-\u3029\u3038-\u303A\uA6E6-\uA6EF]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[\\u16EE-\\u16F0\\u2160-\\u2182\\u2185-\\u2188\\u3007\\u3021-\\u3029\\u3038-\\u303A\\uA6E6-\\uA6EF]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Pc() {
-        var result0;
-        
-        if (/^[_\u203F-\u2040\u2054\uFE33-\uFE34\uFE4D-\uFE4F\uFF3F]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[_\\u203F-\\u2040\\u2054\\uFE33-\\uFE34\\uFE4D-\\uFE4F\\uFF3F]");
-          }
-        }
-        return result0;
-      }
-      
-      function parse_Zs() {
-        var result0;
-        
-        if (/^[ \xA0\u1680\u2000-\u200A\u202F\u205F\u3000]/.test(input.charAt(pos.offset))) {
-          result0 = input.charAt(pos.offset);
-          advance(pos, 1);
-        } else {
-          result0 = null;
-          if (reportFailures === 0) {
-            matchFailed("[ \\xA0\\u1680\\u2000-\\u200A\\u202F\\u205F\\u3000]");
-          }
-        }
-        return result0;
-      }
-      
-      
-      function cleanupExpected(expected) {
-        expected.sort();
-        
-        var lastExpected = null;
-        var cleanExpected = [];
-        for (var i = 0; i < expected.length; i++) {
-          if (expected[i] !== lastExpected) {
-            cleanExpected.push(expected[i]);
-            lastExpected = expected[i];
-          }
-        }
-        return cleanExpected;
-      }
-      
-      
-      
-      
-        var self = this,
-            turo = this.turo || {};
-        var emitter = this.emitter,
-            units = turo.units ||
-                    this.units, // for backwards compat.
-            operators = turo.operators || this.operators,
-            testWriter = turo.writer ||
-                        this.testWriter, // for backwards compat.
-            inputLength = this.inputLength || input.length,
-            parseContext = this.parseContext;
-      
-            
-      
-        this.reservedWords = ["unit", "per", "const", "include", "unit", "let"];
-      
-      
-      var result = parseFunctions[startRule]();
-      
-      /*
-       * The parser is now in one of the following three states:
-       *
-       * 1. The parser successfully parsed the whole input.
-       *
-       *    - |result !== null|
-       *    - |pos.offset === input.length|
-       *    - |rightmostFailuresExpected| may or may not contain something
-       *
-       * 2. The parser successfully parsed only a part of the input.
-       *
-       *    - |result !== null|
-       *    - |pos.offset < input.length|
-       *    - |rightmostFailuresExpected| may or may not contain something
-       *
-       * 3. The parser did not successfully parse any part of the input.
-       *
-       *   - |result === null|
-       *   - |pos.offset === 0|
-       *   - |rightmostFailuresExpected| contains at least one failure
-       *
-       * All code following this comment (including called functions) must
-       * handle these states.
-       */
-      if (result === null || pos.offset !== input.length) {
-        var offset = Math.max(pos.offset, rightmostFailuresPos.offset);
-        var found = offset < input.length ? input.charAt(offset) : null;
-        var errorPosition = pos.offset > rightmostFailuresPos.offset ? pos : rightmostFailuresPos;
-        
-        throw new this.SyntaxError(
-          cleanupExpected(rightmostFailuresExpected),
-          found,
-          offset,
-          errorPosition.line,
-          errorPosition.column
-        );
-      }
-      
-      return result;
-    },
-    
-    /* Returns the parser source code. */
-    toSource: function() { return this._source; }
-  };
-  
-  /* Thrown when a parser encounters a syntax error. */
-  
-  result.SyntaxError = function(expected, found, offset, line, column) {
-    function buildMessage(expected, found) {
-      var expectedHumanized, foundHumanized;
-      
-      switch (expected.length) {
-        case 0:
-          expectedHumanized = "end of input";
-          break;
-        case 1:
-          expectedHumanized = expected[0];
-          break;
-        default:
-          expectedHumanized = expected.slice(0, expected.length - 1).join(", ")
-            + " or "
-            + expected[expected.length - 1];
-      }
-      
-      foundHumanized = found ? quote(found) : "end of input";
-      
-      return "Expected " + expectedHumanized + " but " + foundHumanized + " found.";
-    }
-    
-    this.name = "SyntaxError";
-    this.expected = expected;
-    this.found = found;
-    this.message = buildMessage(expected, found);
-    this.offset = offset;
-    this.line = line;
-    this.column = column;
-  };
-  
-  result.SyntaxError.prototype = Error.prototype;
-  
-  return result;
-})();
