+++ turo/lib/parser/index.js
@@ -1,0 +1,5 @@
+import Parser from './Parser';
+
+export {
+  Parser,
+};
