+++ turo/test/test-async-evaluator.js
@@ -3,11 +3,10 @@ import { test, plan } from 'tap';
 
 import DocumentHelper from '../lib/document/DocumentHelper';
 import { AbstractStorage } from '../lib/storage/abstract-storage';
-import _parser from '../lib/parser';
+import { Parser } from '../lib/parser';
 import { Scope } from '../lib/symbols';
 import { EditableDocument } from '../lib/document';
 
-const { Parser } = _parser;
 ////////////////////////////////////////////////////////////////////////
 var scopeUnderTest;
