+++ turo/test/test-unit-evaluator.js
@@ -1,12 +1,12 @@
 import tap from 'tap';
 import _ from 'underscore';
-import turoParser from '../lib/parser';
+import { Parser } from '../lib/parser';
 import evaluator from '../lib/eval';
 import { Units } from '../lib/units';
 
 const { test, plan } = tap;
 
-const parser = new turoParser.Parser();
+const parser = new Parser();
 parser.scope._unitsTable = new Units();
 
 test("Expressions with units", function (t) {
